refactor(engine): extract the guidelines/CLAUDE.md-injection subsystem to guidelines.py (engine-modularization 8/N)
Move the whole guideline-injection subsystem out of add.py into a new
add_engine/guidelines.py module — the first clean cohesive CLUSTER past the
single-function leaf phase:

  _guideline_block · _inject_block · _rule_file_mode · _strip_inline_block
  _insert_rule_reference · _ensure_claude_reference · _inject_guidelines
  _is_brownfield   (+ the cluster-private _INIT_EXCLUDE constant)
  (add.py 6518 -> 6283)

A transitive-closure AST scan (not just one-level free-names) proved the cluster
is self-contained: it makes ZERO outbound calls to non-cluster add functions. Its
only deps are constants (GUIDELINE_FILES, RULES_FILE_REL, WORKFLOW_HEADINGS,
_GUIDE_BEGIN, _GUIDE_END, _RULE_REF_LINE — all already in constants.py),
_atomic_write (io_state), and stdlib (os/re/sys/Path). None of the 8 are
monkeypatched, so a plain re-export works — no call-site qualification needed
(unlike the identity cluster). The cluster-private _INIT_EXCLUDE (used only by
_is_brownfield) travels with its function.

add.py re-imports the 8 as module globals (cmd_init/cmd_sync_guidelines/
cmd_new_task keep calling bare -> add's re-imported global). The move is provably
verbatim: an AST source-segment diff of all 8 functions (and _INIT_EXCLUDE)
against the pre-move add.py is identical; the add.py diff is +5 (the re-import) /
-240 (the subsystem). The behavior test injects a real `init` and asserts the
marker-delimited ADD block lands in CLAUDE.md byte-identically.

guidelines.py -> {constants, io_state} one-way (imports standalone without add;
no cycle). A cohesive subsystem (AGENTS.md/CLAUDE.md block sync) earns its own
module.

Two-pin model re-aimed (both literals; engine_pin.py never hashes):
  ENGINE_MD5     eae20ce2 -> 6d731aab   (md5 of the shrunk add.py)
  ENGINE_PKG_MD5 7ec748db -> 001ad4a0   (guidelines.py joined the package)

3-tree mirror byte-identical. New test_engine_extract_guidelines.py (identity ·
byte-identical block injection · no-cycle · 3-tree pins). Full suite 1882/0; seam
audit clean (94).

The engine package is now 7 modules (constants · io_state · accessors ·
predicates · identity · guidelines · the entry); add.py 7049 -> 6283 lines across
the 8 extractions. What remains is the cmd_* dispatch core, the UDD validator, and
the report renderer — scoped as their own follow-up tasks.

author: Tin Dang
