+++ add-method/tooling/engine_pin.py
@@ -10,5 +10,5 @@ it can never detect drift. The literal was recorded at the commit that first
 introduced it and is updated only by a deliberate, human-approved task.
 """
 
-ENGINE_MD5 = "eae20ce2e708138a785f6a69bd399e93"  # re-aimed @ extract-io-state (engine-modularization 2/N: 4 IO primitives moved to add_engine/io_state.py, add.py shrank). ENGINE_MD5 is md5(add.py) — re-aimed @ engine-package-skeleton (engine-modularization 1/N: ENGINE_MD5 is now a MANIFEST DIGEST over the package — add.py + add_engine/*.py — computed by the separate engine_manifest.engine_digest(); the 43 public + 6 private constants moved verbatim to add_engine/constants.py, add.py re-exports them; pin stays a literal). prior: re-aimed @ phase-build-guard (audit-hardening, F4: `add.py phase build` is no longer a backdoor around the tests->build gate stack — the `nxt=="build"` guards+snapshots are extracted into a shared `_build_entry(root, state, slug)` helper that BOTH cmd_advance and cmd_phase call, so the admin override runs the IDENTICAL freeze gate · build-expectations · unflagged-freeze + flag stamp · tamper tripwire · §5 scope snapshot; verify's _tamper_guard is now armed on a cmd_phase-entered build; cmd_advance behavior byte-unchanged [pure extraction]; heal loop sets phase=build directly so stays exempt). prior: setup-commit-prompt (cmd_init closes with a one-line tip — both greenfield + brownfield branches — reminding the user to commit the .add/ folder to git so the team shares the ADD state; additive output only, no behavior/gate change). prior: gitignore-bak-seed (.add/.gitignore now lists `pre-update-state.bak.json` — the installer update backup — single-sourced in tooling/templates/gitignore.tmpl kept byte-identical to _GITIGNORE_BODY [parity test]; both installers seed/refresh .add/.gitignore on install+update via _seed_gitignore [pip] / seedGitignore [npm]: seed-if-missing ELSE append-if-absent the missing engine-transient pattern lines, additive-only + idempotent + fail-soft). prior: loose-task-release (release-attribute milestone-free DONE tasks: a "loose task" = milestone-free + done [_task_done] + not in any RELEASES.md `loose tasks:` row; NEW _released_loose_tasks / _releasable_loose_tasks readers [parallel to the milestone track, fail-OPEN]; release_data gains a "loose" record-set + summary["loose"]; _render_releases_row writes a `loose tasks: <slugs|none>` line [kwarg, default none]; cmd_release no-op guard now fires only if NO milestones AND NO loose tasks [a loose-only cut works]; status prints a SEPARATE loose cue; release_security_open floor UNCHANGED). prior: todo-capture (new `add.py todo` backlog verb: `todo "<text>"` captures a jotted idea into state["todos"] [{id=max+1, text, created, status}], bare `todo` LISTS open items, `todo --done <id>` closes one; todo_empty / todo_unknown guards; a todo is NOT a task [no spec/contract/gate] — promote via new-task --fast. additive state key, drift-tolerant). prior: standalone-fast-task (blessed milestone-free fast lane: cmd_new_task AFFIRMS a --fast task created with no owning milestone — "standalone fast task ... milestone-free by design" — instead of the intake nudge; cmd_check gains an ADDITIVE `infos`/`informed` tier so a fast standalone is a soft INFO, NEVER a WARN/orphan; the non-fast orphan path is UNCHANGED [test_v8_1_orphan_guard green]). prior: multirepo-federation (`federate pull <id>` lands a producer repo's published immutable contract snapshot at the local .add/contracts/<id>.json [the cross-REPO transport]; `_federation` reader; fail-loud unknown/missing/invalid/version-mismatch; cmd_check federation_source_unreadable WARN; opt-in byte-identical). prior: cross-component-milestone (cmd_advance scenarios->contract HOLD: a `consumes:` task of a DECLARED contract is refused producer_contract_unfrozen until the producer snapshot exists [the intra-milestone BE->FE ordering]; opt-in byte-identical). prior: cross-component-contract+refute-heal (null-hash snapshot now HARD-STOPs (fail-loud Finding 2); a corrupt LIVE snapshot now surfaces contract_snapshot_unreadable WARN not silent no-stale (Finding 3). NEW _contracts/_task_produces/_task_consumes/_contract_snapshot/_contract_body_hash/_contract_findings + produces:/consumes: header regexes; cmd_advance contract->tests crossing WRITES .add/contracts/<id>.json on a `produces:` freeze [idempotent, atomic] + PINS the live hash on a `consumes:` cross [missing snapshot HARD-STOPs contract_snapshot_missing]; cmd_check adds contract_producer_unknown RED + contract_consumer_stale WARN. Hash is over the §3 fenced SHAPE so a pure version bump does not churn. Opt-in byte-identical. prior: per-component-verify v3 (NEW pure _cite_region: the Build-expectations evidence region [matches the standard ### heading AND the fast-lane bare line, up to GATE RECORD] with the engine stamp STRIPPED — closes refute-read #3 Finding 1 [stamp landing inside the block self-satisfies] + Finding 2 [fast-lane template has no ### heading -> component-bound fast task permanently false-blocked]; cmd_gate cite-check now calls _cite_region; prior v2 (cite-search SCOPED to the §6 "### Build expectations" block — subsumes the v1 stamp-strip AND closes refute-read Finding 2 [a generic green_bar like "all tests pass" that collides with §6 checklist/Outcome boilerplate no longer self-passes]; prior: refute-heal (cite-check now STRIPS the engine's own "·expected green-bar:" stamp before the §6 substring search, so a prior HARD-STOP's stamp can't self-satisfy a later PASS [BLOCKER from refute-read]; component-aware-add M2 per-component-verify: NEW pure _task_green_bar; cmd_gate gains a completing-only cite-gate (component_green_bar_uncited) for a bound task with a green_bar — surfaces + stamps "component: X · expected green-bar: B" into §6 via _stamp_gate_record; cmd_check WARNs component_green_bar_unset. Engine still never runs a suite. Unbound/no-green_bar = byte-identical (every branch guarded behind _task_green_bar is not None; HARD-STOP exempt). risk:high. red suite test_per_component_verify.py [9]). PRIOR ⤵ re-aimed @ component-registry [self-heal: degrade-safe tasks/ scan in _component_findings + reserve "?" component name — refute-read MAJOR+MINOR] (component-aware-add M1 component-registry: NEW component registry — readers _components/_component_root/_task_component/_component_findings + _COMPONENT_LINE_RE; _declared_scope EXTENDED additively to append a bound component's root (compose, never redraw explicit-token resolution); cmd_check gains fail-closed RED registry findings (components_malformed/component_unknown/component_root_outside, like wave_ledger_malformed); TASK.md.tmpl gains a commented `component:` affordance; guarded `import tomllib`. OPT-IN — no .add/components.toml ⇒ byte-identical to pre-component ADD. risk:high. red suite test_component_registry.py [18]). PRIOR ⤵ re-aimed @ loop-aware-orient (status+guide STEER into the loop at the loop juncture. NEW pure/fail-soft fn _done_resume(root,state,slug)->(headline,next_step,chapter): classifies a DONE task from its milestone exit-criteria tally — LOOP-JUNCTURE total>0&met<total -> name the unmet goal + route 09-the-loop.md; GOAL-MET total>0&met==total -> milestone-done; PLAIN (no criteria/no milestone/read error) -> today's "start the next feature"/02-the-flow.md, BYTE-IDENTICAL. cmd_status done-branch + cmd_guide human AND --json done-branch route via the helper, overriding only when chapter!=02-the-flow.md so PLAIN output is unchanged. loop.md cue HEALED: the engine now prints "goal not met (m/n exit criteria)" so the doc's status-attribution is TRUE (no doc edit needed). red suite test_loop_aware_orient.py [9]. risk:low — additive + fail-soft). PRIOR ⤵ re-aimed @ ccsk-rule-file (re-integration of fork PR #45 onto post-1.9.0 main: ccsk-aware rule-file placement. NEW module consts RULES_FILE_REL=.claude/rules/add-workflows.md + WORKFLOW_HEADINGS + _RULE_REF_LINE; NEW pure/fail-soft fns _rule_file_mode (disk-derived: --rule-file OR a .ccsk/ dir OR an existing rule file — no persisted state) + _strip_inline_block (migrate an inline ADD:BEGIN..END out; an unterminated BEGIN is left as-is) + _insert_rule_reference (bullet under an existing Workflows/Rules heading at any # level, else append a fresh ## Workflows) + _ensure_claude_reference (strip inline + add the ref; .bak on change; idempotent). _inject_guidelines gains a rule_file param + a CLAUDE-only branch: the full block goes to .claude/rules/add-workflows.md and CLAUDE.md keeps a reference bullet — AGENTS.md stays inline. cmd_init + cmd_sync_guidelines pass getattr(args,'rule_file',False); a --rule-file store_true is added to the init + sync-guidelines parsers. WITHOUT the flag / a .ccsk dir / an existing rule file: output+scaffold BYTE-IDENTICAL to today. red suite test_rule_file_mode.py [11] + test_installer_prompts.py [+4 pip/npm parity]. risk:low — additive + fail-soft. credit: original feature by Crystal D. (@imnortheastt)). PRIOR ⤵ re-aimed @ fast-new-task-flag (fast-lane 3/4: `add.py new-task <slug> --fast` opts into the fast lane. cmd_new_task renders the MINIMAL `_render_template("TASK.fast.md", …)` (vs "TASK.md") + records the durable marker `state["tasks"][slug]["fast"] = True` (ABSENT on a normal task — grandfathered). The `new-task` parser gains a `--fast` store_true. cmd_status tags an ACTIVE fast task with a ` · fast` marker on the `active  :` line (presentation-only, existence-gated). v1→v2 CHANGE-REQUEST (human-chosen "fast implies floor" at the re-freeze): v1's §3 wrongly claimed the freeze floor was "enforced upstream, UNCONDITIONAL, no new guard" — FALSE, since the existing `unflagged_freeze` check only fires once §3 is ALREADY frozen, so a DRAFT-§3 task reached gate=PASS unrefused. v2 WIRES `--fast` into the freeze-before-build gate (shipped by task 2): the gate's freeze-firing condition becomes `_freeze_gated := _optin OR state["tasks"][slug].get("fast") is True`, so a fast task CANNOT cross tests→build with a DRAFT §3 under ANY milestone (opted-in or not) — collapse-never-skip is REAL for the fast lane itself. The build-expectations gate stays opt-in-only (`_optin`). `--fast` composes with `--from-delta` (both pre-fill §1 Feature; the fast template has it). WITHOUT `--fast`: output + scaffold + state BYTE-IDENTICAL to today (renders TASK.md, no `fast` key) — zero ripple, no existing task carries the marker. risk:high · §3 FROZEN @ v2 (approved by Tin Dang). red suite test_fast_new_task_flag.py [8 tests: minimal-scaffold+marker · plain-new-task-unchanged · fast floor holds under a PLAIN milestone (contract_not_frozen) · plain task NOT freeze-gated (zero ripple) · fast completes through gates to PASS+§6 stamped · status marks/omits the fast tag · --fast ∘ --from-delta]. add.py logic+parser delta → ENGINE_MD5 80985fc3 → d4807ff9, re-synced across all 3 copies; full suite green. PRIOR re-aim @ freeze-before-build-gate (fast-lane 2/4: makes "collapse-never-skip" REAL — a NEW opt-in freeze-before-build gate in cmd_advance at the `nxt == "build"` crossing, placed BEFORE the existing build-expectations gate. GUARD: `_optin := task has a milestone M AND state["milestones"][M]["await_confirm"] is True`; `if _optin and not _contract_frozen(_raw_phase_bodies(root, slug).get(3, "")): _die("contract_not_frozen")`. Mirrors the build-expectations gate's opt-in switch EXACTLY (the same --await-confirm marker, one level out) — a plain / no-await_confirm / milestone-less task is NEVER gated, so every existing advance-to-build flow stays byte-green (ZERO ripple, MEASURED). Refactor: `_ms`/`_optin`/`raw3` computed ONCE at the build crossing and reused by both opted-in gates (freeze, then build-expectations) + the existing unflagged_freeze flag check. validate-then-write — the refusal runs BEFORE the tripwire/scope snapshots, writes nothing, task stays at `tests`. EMPIRICAL JUSTIFICATION: the engine did NOT hard-enforce freeze-before-gate for ANY task before this — a DRAFT §3 task reached gate=PASS unrefused (the freeze was a skill/human convention only); an ALL-tasks guard breaks ~40 fixture files (81 fails at build-entry / 176 at gate — the confirm-parent trap), so the human chose the opt-in path. SIBLING-FIXTURE FIX (the ONLY measured ripple): test_build_expectations_gate's `_optedin_task_at_tests` helper now freezes §3 (+ a well-formed least-sure flag) so its 2 opted-in fixtures clear the new gate and still reach the build-expectations check. red suite test_freeze_before_build_gate.py [6 tests: opted-in unfrozen→contract_not_frozen + stays tests + no scope-snapshot · opted-in frozen+filled→advances · plain milestone unfrozen→advances · no-milestone unfrozen→advances · freeze PRECEDES build-expectations (DRAFT §3 + placeholder §6 → contract_not_frozen not build_expectations_unfilled) · frozen §3 + placeholder §6 → build_expectations_unfilled takes over]. risk:high · §3 FROZEN @ v1 (approved by Tin Dang). add.py logic delta → ENGINE_MD5 280daae8 → 80985fc3, re-synced across all 3 copies; full suite green. PRIOR re-aim @ fast-lane-template (fast-lane 1/3: NEW `_FALLBACK_TASK_FAST` minimal-template circuit-breaker analog [sections {0,1,3,4,5,6}; §2 SCENARIOS + §7 OBSERVE dropped] + `_render_template` fallback branch extended to "TASK.fast.md" via a name→fallback map [TASK.md + TASK.fast.md], so a deleted templates/ never hard-fails the fast lane. Ships templates/TASK.fast.md.tmpl ×3 byte-identical [canonical + dogfood + _bundled]; the full TASK.md.tmpl is byte-untouched; the --fast flag + cmd_new_task wiring + check/audit tolerance are the NEXT task [fast-new-task-flag]. The minimal template keeps the trust floor: §3 freeze-flag + Status [_flag_well_formed], §6 GATE RECORD/Outcome/Reviewed [_stamp_gate_record], §0 Anchors [_grounded_state], §4 red-before-build, §5 Scope. risk:high · conservative gate · §3 FROZEN @ v1 (approved by Tin Dang). red suite test_fast_lane_template.py [25 tests]. add.py delta → ENGINE_MD5 cb7ddd03 → 280daae8, re-synced across all 3 copies; full suite green. PRIOR re-aim @ gate-record-writeback (flow-enforcement 3/3 — FINAL: the gate write-back — `add.py gate <outcome>` now mirrors the resolved verdict into the task's §6 `### GATE RECORD`, so the file and state.json never silently diverge (Finding C: `gate PASS` recorded state but left the §6 a `<…>` placeholder, which `audit` flags `malformed_gate_record`). NEW `_stamp_gate_record(root, state, slug, outcome)` called in cmd_gate AFTER save_state (state is the source of truth; the file only mirrors it): rewrites the §6 `Outcome:`/`Reviewed by:` lines (+ the `If RISK-ACCEPTED -> owner·ticket·expires` line from the state waiver) to resolved values via `_actor_stamp(state).name` + `date.today()`. NO opt-in — UNLIKE the two refusal gates (tasks 1&2), a write is ADDITIVE and never refuses, so it cannot break the census; the SAFETY is GRANDFATHER: each line is rewritten ONLY while it still holds a `<…>` placeholder, so a hand-filled/already-resolved record is byte-untouched, and a re-gate never clobbers. No GATE RECORD block / no placeholder / unreadable file → silent no-op, the file stays byte-identical (a completing gate still syncs the `phase:` marker as before — orthogonal). The audit divergence detector (`malformed_gate_record` ≠1 Outcome line · `gate_record_mismatch` §6≠state) was ALREADY present — the write-back is what finally makes opted-in-or-not tasks PASS it, project-wide. §3 FROZEN @ v1 (all-tasks, no opt-in — human-chosen at the freeze over the opt-in draft). risk:high · conservative gate. red suite test_gate_record_writeback.py [6 tests]. Build correction (faithfulness, surfaced at build): the "no §6 block → file byte-unchanged" scenario over-claimed — a completing `gate` ALWAYS rewrites the `phase:` marker (verify→done) independent of write-back; the test normalizes that one line and additionally asserts no GATE RECORD block was fabricated. Prose lint: the `_stamp_gate_record` docstring's "two refusal seams" tripped the ubiquitous-language guard (seam = a Group-C machine constant, banned as prose) → reworded "gates". add.py logic delta → ENGINE_MD5 d7a104fa → cb7ddd03, re-synced across all 3 copies; full suite green. PRIOR re-aim @ build-expectations-gate (flow-enforcement 2/3: the tests->build verify-expectations gate — `cmd_advance` (build-entry) now refuses `build_expectations_unfilled` when an OPTED-IN task's §6 `### Build expectations` block is a `<…>` placeholder/empty, BEFORE the tripwire/scope snapshots (validate-then-write — a refused advance leaves phase=tests, no sidecar). OPT-IN = the task's PARENT milestone has a `confirmed` key (the SAME --await-confirm switch task-1 uses, one level out) — so `--await-confirm` now arms the WHOLE flow-enforcement suite (content-aware confirm + build-expectations) per milestone, and the 33 existing advance-to-build suites (plain/no milestone) stay green = ZERO ripple. `_section_unfilled` EXTENDED compatibly: break the section at ANY next header (`#`, not just `## `) + SKIP `>` blockquote guidance lines, so it reads the `### Build expectations` SUB-section; task-1's `## Shared / risky contracts` truth table is preserved (test_contract_fill_gate stays green). risk:high · conservative gate · §3 FROZEN @ v1 (milestone-inherited opt-in, approved by Tin Dang). red suite test_build_expectations_gate.py [6 tests]. OPT-IN MARKER (v1.1, caught by the full suite at build): keying the gate on `"confirmed" in m` broke the census — `milestone-confirm` stamps confirmed:true on a PLAIN milestone too, so at advance time a never-opted-in milestone looked opted-in. Fix: `new-milestone --await-confirm` now ALSO seeds a stable `await_confirm:true` marker (set ONLY at creation, never by milestone-confirm), and the build gate keys on `milestone.await_confirm is True`. add.py logic delta → ENGINE_MD5 428ca1d1 → ee43743c → d7a104fa, re-synced across all 3 copies; full suite green. PRIOR re-aim @ contract-fill-gate (flow-enforcement 1/3: the CONTENT-aware confirm gate — `milestone-confirm` now refuses `milestone_contracts_unfilled` when an OPTED-IN milestone's MILESTONE.md `## Shared / risky contracts` section is still a `<…>` placeholder/empty, so "confirmed" MEANS the cross-task contracts were present. NEW pure `_section_unfilled(md_text, header)` [PRESENT-but-(empty|`<…>`)→True; ABSENT→False grandfather; filled→False] — the shared placeholder predicate the sibling `build-expectations-gate` reuses for §6. Gate is OPTED-IN ONLY (`"confirmed" in m`): a grandfathered no-key milestone keeps the plain stamp, so the census + every existing flow stays byte-green. CHANGE-REQUEST v1→v2 at BUILD: v1 gated ALL milestones → the new-milestone scaffold ships the placeholder, so `test_min_pillar`'s census `milestone-confirm mvp` (a no-key scaffold) broke; v2 scoped the gate to opted-in milestones, containing the ripple to test_confirm_parent's 2 opted-in tests (now fill contracts first — in §5 scope). validate-then-write — refuse BEFORE the confirmed:true write. risk:high · conservative gate · §3 FROZEN @ v2 (approved by Tin Dang). red suite test_contract_fill_gate.py [7 tests]. add.py logic delta → ENGINE_MD5 9258dcc7 → 428ca1d1, re-synced across all 3 copies; full suite green. PRIOR re-aim @ confirm-parent (lean-pass M3 flow-simplification 2/3: the OPT-IN confirm-parent gate — `new-milestone --await-confirm` seeds the milestone confirmed:false; WITHOUT the flag NO `confirmed` key is written → grandfathered-confirmed → no gate (so the 342 existing engine tests stay byte-green untouched, ZERO churn); `cmd_new_task` HARD-BLOCKS (validate-then-write, before any scaffold) on `not _milestone_confirmed(state, milestone)` → `_die("milestone_unconfirmed: confirm it first — add.py milestone-confirm <slug>")`; `add.py milestone-confirm <slug> [--by]` is the human gate (validate-then-write — unknown→unknown_milestone; sets confirmed:true/at/by/actor; re-confirm idempotent note; engine RECORDS, never self-confirms). NEW pure `_milestone_confirmed(state, mslug)` mirrors `_setup_locked` one level down (True if confirmed is True OR key absent; False only if confirmed is False). cmd_new_milestone seeds the confirmed/confirmed_at/confirmed_by trio ONLY under `--await-confirm` (plain-create record + output byte-unchanged); `--await-confirm` flag + `milestone-confirm` subparser added. The v1 DEFAULT-ON design (every new milestone gated) was REJECTED at the v1→v2 change-request: its true cost was 342 broken test methods across 61 files (mixed call-shapes, not cleanly scriptable), ~7× the freeze estimate, for no gain over opt-in (the process lesson is about the GUIDED draft-then-confirm flow, where the skill passes the flag). risk:high · conservative gate · §3 FROZEN @ v2 (opt-in, approved by Tin Dang). red suite test_confirm_parent.py [10 tests]. add.py logic+parser delta → ENGINE_MD5 3f82050f → 9258dcc7, re-synced across all 3 copies; full suite green (a verify-gate adversarial refute-read also corrected one stale "(default-on)" inline comment in cmd_new_task → "(OPT-IN)", comment-only, no behavior change). PRIOR re-aim @ identity-in-status (user-identity foundation 3/3 — FINAL, team-collaboration milestone 2: SURFACE the git-native actor — READ-ONLY presentation over task-1's resolver + task-2's stamps, no new writes. (1) cmd_status human gains an `actor   :` line (after `project autonomy:`) = `_whoami(state)` rendered `<name> [<email>] (source: <source>)`, ALWAYS present (resolver TOTAL). (2) `status --json` gains an `"actor"` key = `_whoami(state)` [ratified into test_wave_status_hint's frozen sanctioned set, mirror of active_milestones/active_tasks]. (3) `report_data` task rows gain `gate_actor` [t.get, None when unstamped] + the milestone dict gains `done_actor` — the JSON report facts. (4) `render_report` shows the recorded actor present-only: a ` closed by <name>` header line when done_actor is set + a ` GATED BY` block listing each task's gate_actor; NEW `_fmt_actor(actor)` helper [name <email>, "" when absent]. DESCRIPTIVE + READ-ONLY: no save_state, no new state key, no decision change; an unstamped/legacy record renders NO actor (never a placeholder/crash). FREEZE seam still excluded (carried §7 delta from actor-stamping). risk:high · conservative gate · §3 FROZEN @ v1 (auto-mode standing authorization). red suite test_identity_in_status.py [5 tests] + the sanctioned-set census ratification. add.py logic delta → ENGINE_MD5 d8b9c699 → 04c83b4a, re-synced across all 3 copies; full suite green. PRIOR re-aim @ actor-stamping (user-identity foundation 2/3, team-collaboration milestone 2: STAMP the structured actor on the FOUR engine-WRITTEN human seams, ALONGSIDE today's free-text. NEW `_actor_stamp(state)` = `_whoami(state)` [the SINGLE source; TOTAL — always a non-empty name, so a stamp never fails/blocks a seam] + `_render_actor_line(state)` [`<name> [<email>] (<source>)` for the ledger row]. SEAMS: cmd_lock writes `state["setup"]["actor"]` alongside the existing locked_by free-text · cmd_gate writes `state["tasks"][slug]["gate_actor"]` on EVERY recorded outcome (PASS·RISK-ACCEPTED·HARD-STOP — who recorded the verdict) · cmd_milestone_done writes `state["milestones"][slug]["done_actor"]` · cmd_release adds an `actor:` line to the RELEASES.md row via `_render_releases_row(...,actor=_render_actor_line(state))` [still NO state.json write]. DESCRIPTIVE + ADDITIVE: new optional fields, absent on legacy records → behaves as today; NO existing command's decision reads the stamp; the actor field rides the EXISTING save_state/_atomic_write paths (no new IO, no new failure mode). FREEZE seam EXCLUDED (no engine write — human-authored TASK.md text; seeded as a §7 delta for a future `add.py freeze` command — human-confirmed). risk:high · conservative gate · §3 FROZEN @ v1 (auto-mode standing authorization). red suite test_actor_stamping.py [9 tests]. add.py logic delta → ENGINE_MD5 6f28abab → 610fae59, re-synced across all 3 copies; full suite green. PRIOR re-aim @ actor-identity (user-identity foundation 1/3, team-collaboration milestone 2: the git-native ACTOR resolver — the engine's FIRST git call, strictly fail-soft. NEW `_git_config(key)` [shutil.which gate · subprocess.run(["git","config","--get",key], timeout=2) · catch OSError/SubprocessError/ValueError · empty→None — NEVER raises/hangs, fixed argv, no shell] + `_os_user()` [getpass.getuser() floored, catches KeyError/OSError → "unknown" sentinel] + `_whoami(state)->{name,email,source}` [priority: actor_override (non-blank name)→'override' · git config user.name/email→'git' · _os_user()→'os' floor; TOTAL, name always non-empty, email may be None] + `cmd_whoami` [no-flags show `actor : <name> [<email>] (source: …)` + --json · --name/--email store actor_override (email ""→None) · --unset clear; validate-before-write: reject no_actor_override · actor_name_blank] + the `whoami` subparser (--name/--unset mutually exclusive at the parser → argparse exit 2) + imports shutil/subprocess. NEW optional state key `actor_override`={name,email} (absent by default → behaves as today). DESCRIPTIVE only, no existing command's decision changes. risk:high · conservative gate · §3 FROZEN @ v1 (auto-mode standing authorization). LAYERED ON the fold-suggestion-seams base (1911b250) after a live parallel-writer contention reconciliation (the fold session committed feat/fold-suggestion-seams @ 9dc3e944 = PR #47, clobbering the in-flight actor-identity engine in the shared tree; re-applied additively here). A verify-gate independent review [python-expert, BLOCK, 0.97] had found TWO real crash paths violating the frozen fail-soft/TOTAL invariants, both fixed before this re-application: (1) `_git_config` now catches UnicodeDecodeError [a ValueError subclass — a non-UTF-8/latin-1 git config value would otherwise crash the engine's first git call]; (2) `_os_user` floors getpass.getuser()'s KeyError [bare container: no passwd entry + no LOGNAME/USER] → "unknown" so _whoami stays TOTAL. Plus 2 NITs: --email "" normalizes to None; --name/--unset mutually exclusive. red suite test_actor_identity.py [17 tests: + test_failsoft_nonutf8_value · test_os_floor_total_when_getuser_raises · test_set_empty_email_normalized_to_none · test_name_and_unset_mutually_exclusive; test_failsoft_nonzero_or_empty → test_failsoft_empty_stdout] + test_min_pillar census (LIFECYCLE += `whoami`). add.py logic+parser delta → ENGINE_MD5 1911b250 → 6f28abab, re-synced across all 3 copies; full suite green. PRIOR re-aim @ parallel-status-view (team-collaboration foundation 4/5: cmd_status renders the active SET as PARALLEL STREAMS — presentation-only, no command DECISION changes. THREE additive edits: (1) the human-readable branch prints a `streams : <N> active milestones` block AFTER the unchanged `active  :` line, gated on len(active_milestones)>=2, one line per active milestone [PRIMARY first via [_primary]+[m!=_primary], primary marked `▸` + tagged `(primary)`] naming task=active_tasks[m] or (none) and phase=tasks[tk].phase or `-`; (2) the milestone-rollup mark becomes `"*" if mslug in active_milestones else " "` [N<=1 identical since active_milestones==[primary] or []; N>=2 marks every active member]; (3) `status --json` gains additive keys active_milestones [list] + active_tasks [map], every existing key UNCHANGED. N<=1 text + existing json byte-for-decision identical — the full suite is the untouched oracle [the standing additive-cue convention shared with the graduation/release/spec/fold cues]. risk:high · conservative gate · §3 FROZEN @ v1 (auto-mode standing authorization). add.py logic delta → ENGINE_MD5 929ced7e → fa8e9818, re-synced across all 3 copies; red suite test_parallel_status_view.py [8 tests], full suite green. PRIOR re-aim @ multi-active-commands (team-collaboration foundation 3/5: the multi-active COMMANDS that USE the task-2 seam — NEW _activate_milestone(state, slug) [append-to-SET-if-absent + focus the scalar mirror + repoint scalar active_task from active_tasks[slug]] · _deactivate_milestone(state, slug) [remove-from-SET + pop active_tasks[slug]; if it was the primary, repoint to the SET's last member, else leave the primary untouched]. NEW verbs: `activate <slug>` [cmd_activate: reject unknown_milestone if slug∉milestones, milestone_done if status==done, then _activate_milestone] · `deactivate <slug>` [cmd_deactivate: reject milestone_not_active if slug∉active_milestones, then _deactivate_milestone] — both validate-before-mutate (a reject leaves state byte-unchanged). cmd_use made MILESTONE-AWARE: focuses the task's OWN milestone (_activate_milestone(ms) + _set_active_task(state, slug, ms)) when the task carries a known milestone, else scalar-only (back-compat) — closing the task-2 nit-2 §7 delta (cross-milestone map entry). cmd_archive_milestone's clear-pair replaced by _deactivate_milestone(state, slug) (drop-from-SET + pop the task entry, repointing primary). new-milestone keeps its REPLACE-to-focus (additive activate/deactivate are the SET writers; task-2's frozen _set_active_milestone untouched). risk:high · conservative gate · §3 FROZEN @ v1 (auto-mode standing authorization). add.py logic+parser delta → ENGINE_MD5 5eb79fa8 → 458fb992 → 929ced7e, re-synced across all 3 copies; red suite test_multi_active_commands.py [13 tests], full suite green. v2 (929ced7e) = a verify-gate independent review [python-expert, BLOCK, 0.97] found a NON-primary-archive N≤1 regression: routing cmd_archive_milestone solely through _deactivate_milestone DROPPED the task-2 guard that cleared the scalar active_task when the archived milestone's MEMBER task was the scalar focus while the PRIMARY was a different milestone (reachable via new-milestone replace-to-focus, which leaves the scalar stale) — leaving active_task dangling at a now-deleted task (phantom unknown_task for slug-less advance/gate). Fixed by restoring the N≤1 guard AFTER the SET-level deactivate: `if _active_task(state) in members: state["active_task"]=None` (the map entry is already popped by _deactivate_milestone; the frozen archive-routes-through-_deactivate clause stands, the back-compat scalar clear is additive) + a red→green guard test_archive_nonprimary_clears_stale_scalar_task. Review NOTE (not blocking, contract-faithful): cmd_use re-activates a DONE milestone (no done-check, unlike cmd_activate) → seeded as a §7 delta. PRIOR re-aim @ active-accessors (team-collaboration foundation 2/5: the active milestone/task ACCESSOR SEAM — NEW _active_milestone(state) · _active_task(state, milestone=None) · _set_active_milestone(state, slug) · _set_active_task(state, slug, milestone=None) [reads return the N≤1 scalar mirror; writes keep scalar + active_milestones/active_tasks in sync; orphan active-task = scalar-only; clear pops the map entry] + EVERY active-milestone/active-task read & write site routed through them (incl. cmd_check's formerly-direct read — consumes the state-schema-migration §7 deltas: the 3rd seam + active_tasks partial-state default). Behavior-preserving for N≤1: the full suite stays green with NO test changed (the byte-for-decision-identity oracle). risk:high · conservative gate. §3 FROZEN @ v1 (auto-mode standing authorization; [contract] "primary"=scalar today, per-milestone arg exists). add.py logic delta → ENGINE_MD5 aef59d9c → 995d9369 → 5eb79fa8 (v2: a verify-gate independent review [python-expert, MERGE-WITH-NITS, 0 blocking, 0.97] found a stale active_tasks entry left after archive — fixed by clearing the active task with its milestone still known [_set_active_task(state, None, slug) before _set_active_milestone(state, None)] so active_tasks[slug] is popped; nit 2 [cmd_use cross-milestone map entry, pre-existing] → §7 delta for multi-active-commands), re-synced across all 3 copies; red suite test_active_accessors.py [10 tests], full suite green) · re-aimed @ state-schema-migration (team-collaboration foundation 1/5: the multi-active state schema + a PURE/idempotent/TOTAL `_migrate_state(state)` forward-migration wired into BOTH load seams — load_state + _load_state_for_json — that grandfathers a single-active state to the new shape on load: adds `active_milestones` [the active SET, derived from the scalar active_milestone: None→[], "x"→["x"]] + `active_tasks` [per-milestone active-task MAP]; KEEPS the scalar active_milestone/active_task as the N≤1 mirror so the ~20 not-yet-routed readers keep working; an orphan active_task [one belonging to no active milestone] stays a top-level scalar fallback [FROZEN decision (a)]; an already-migrated state is returned unchanged [idempotent]; corrupt JSON still dies state_invalid in the loader [migration never reached]. init is BORN-MIGRATED [active_milestones:[] + active_tasks:{} written natively]. risk:high · conservative gate. §3 FROZEN @ v1 ([contract] orphan=top-level fallback · storage=active_tasks map · scalar mirrors kept). add.py logic delta → ENGINE_MD5 25c6fcc5 → 8a914147 → aef59d9c (v2: a verify-gate independent review [python-expert, MERGE-WITH-NITS, 0 blocking] flagged the "PURE" docstring/contract word against the in-place mutation — fixed so the helper upgrades on a fresh top-level `dict(state)` copy and NEVER mutates the caller's dict; behavior otherwise identical, 11 tests stay green), re-synced across all 3 copies; red suite test_multi_active_state.py [11 tests], full suite green) · re-aimed @ scope-walk-artifacts (exclude gitignored build artifacts — `.next`/`coverage`/`test-results` dirs + `.pyc`/`.tsbuildinfo` suffixes — from the §5 scope-walk, alongside the existing `.serena` cache; supersedes the fold-command pin) · re-aimed @ fold-command (delta-resolution 4/4: NEW `add.py fold` — the HUMAN-AUTHORIZED reversal of the frozen "engine stays judgment-free; there is no add.py fold" principle (test_foundation_update_loop re-frozen @ v3). Mechanizes ONE competency-lesson consolidation session, TRANSCRIPTION-ONLY: it routes each lesson's OWN captured text into its routed home and NEVER composes/merges prose (consolidation judgment stays the human's, via the compaction door). cmd_fold [--task/--comp] + the `fold` subparser + PURE _fold_competency_delta(text, version, comps) [flips EVERY open `[<COMP> · open]` → `folded` + ` [folded foundation-version N]` stamp via _COMP_OPEN_TOKEN_RE, mirror of _resolve_spec_delta; line text + `(evidence: …)` byte-preserved; None when nothing open] + _FOLD_ROUTES {DDD/SDD/UDD → PROJECT.md §Domain/§Spec/§Users · TDD/ADD → CONVENTIONS.md §Method learnings, TOTAL over the five} + _section_present/_prepend_to_section/_prepend_key_decision_row. BATCH = one call → ONE foundation-version bump → every stamp carries the SAME N (fold.md "one session, one bump"). VALIDATE-ALL-THEN-WRITE: ≥1 open lesson in scope · every routed section present (incl. §Key Decisions) · parseable `foundation-version:` — all checked + every new body built in memory BEFORE any _atomic_write, so a reject leaves the WHOLE tree byte-unchanged. 3 reject codes: no_open_deltas · missing_route_section · no_foundation_version. `fold`/`folded` are Group C machine tokens referenced by NAME (_FOLD_VERB/_FOLDED) inside output strings so the ubiquitous-language prose lint sees no slang. AUTHORIZED guard change-requests: test_foundation_update_loop @v2 → @v3 (invert no-fold-command → the command EXISTS + transcription-only; REJECT_CODES → the new 3) · test_min_pillar census (LIFECYCLE += `fold` · _NONZERO_OK += `fold` — refuses no_open_deltas on a delta-free board) · test_ubiquitous_language MACHINE_CONSTANTS += `fold`. fold.md rewritten ×3 to describe the command (prose stays "consolidation"; `fold`/`folded` in code spans only). risk:high · conservative gate. §3 FROZEN @ v1 ([contract] transcribe-not-consolidate + uniform `(<COMP>)`-prefixed verbatim bullet, human-approved at the freeze). VERIFY-gate refute-read hardened the cross-file commit: cmd_fold now stages EVERY edit to a temp then renames all via the NEW _atomic_write_many(writes) two-phase helper, so a mid-commit IO failure (disk-full/permission) leaves NOTHING written — the frozen §3 "any failure → write nothing" now holds across the N task-files + PROJECT.md + CONVENTIONS.md set, not just per-file; writes ordered foundation-FIRST so the near-impossible mid-rename residual is a still-open lesson (visible, re-runnable) never a flipped-but-untranscribed one (silent loss). Success + precondition-reject paths byte-identical (7 fold tests stay green). A SECOND VERIFY-gate finding (dogfooding HARD-STOP, human-authorized to fold into THIS task): the §5 scope guard's `_scope_walk` excluded only `.git/.add/__pycache__/node_modules` — NOT `.serena`, serena's code-intelligence symbol-index cache, which re-writes itself whenever a source file changes. So the AI's OWN add.py edits churned `.serena/cache/*.pkl`, the build-entry snapshot baked them in, and the gate flagged them as out-of-scope touches (un-fixable at gate time: deletion reads as a delete-touch, regeneration as a modify-touch) — exhausting the heal loop to a FALSE-POSITIVE HARD-STOP. Fix: `.serena` added to `_SCOPE_EXCLUDE_DIRS` (a tool cache with no build signal, the exclude set's exact purpose; widening it is the sanctioned additive change-request the constant's own comment names). Proven red/green: pre-fix `.serena/...pkl` leaks into the snapshot, post-fix pruned — guarded by a new assertion + fixture in test_scope_gate_enforce.py (added to §5 scope). The task was reopened → re-crossed tests→build so the FRESH snapshot excludes `.serena` from birth, making the gate immune to serena re-indexing. A pre-MERGE independent review (python-expert, verdict MERGE-WITH-NITS, zero blocking) flagged ONE faithfulness nit: the cmd_fold inline commit comment overstated atomicity ("leaves NOTHING written" is true only for the phase-1 temp-write failure; a phase-2 mid-rename residual re-runs by DUPLICATING, not cleanly). Comment-only honesty fix — no behavior change (the two-phase commit + foundation-first ordering stand). add.py logic+parser delta → ENGINE_MD5 392b4e55 → 58ad75cb → b65fc266 → f78a2f24 → 5f3f7d28, re-synced across all 3 copies; red suite test_fold_command.py [7 tests] + test_scope_gate_enforce .serena guard, full suite green). Supersedes re-aimed @ spec-delta-guards (delta-resolution 3/4: SPEC-delta GUARDS make an open SPEC delta impossible to lose silently — PROJECT-WIDE (human change-request from member-scoped at the freeze; report unified too). cmd_compact gains a `open_spec_deltas_unresolved` HARD guard AFTER open_deltas_unfolded, BEFORE the first rename (validate-before-move; a reject leaves tree+state byte-unchanged) — refuses while ANY task in the project holds an open SPEC delta, deliberately broader than the member-scoped competency guard (a SPEC delta is a project-global forward hand-off). Three READ-ONLY nudges, all on the SAME source `len(_collect_open_spec_deltas(root))`: cmd_status `spec : N open SPEC delta(s) — resolve: new-task --from-delta / drop-delta` (sibling of the `deltas : N open` fold nudge) · cmd_milestone_done `note: N open SPEC delta(s) to resolve (seed/drop)` (sibling of the competency fold note) · report_data.summary[open_spec] + a render ` SPEC DELTAS  N open` line. NO new collector (reuses the task-1 reader), NO new subcommand (census untouched — task-2 verb-vs-flag lesson). add.py logic delta → ENGINE_MD5 5120b0c2 → 392b4e55, re-synced across all 3 copies; red suite test_spec_delta_guards.py [5 tests], full suite green. Supersedes re-aimed @ seed-and-drop (delta-resolution 2/4: NEW SPEC-delta resolution VERBS — the first WRITERS of the seeded/dropped statuses task 1 only TOLERATED on read. (1) `new-task <new> --from-delta <prior>` SEEDS the prior's first open SPEC delta into the new task — reads it via _first_open_spec_text, pre-fills §1 `Feature: <text> (from <prior> spec-delta)`, flips the source `[SPEC · open]` → `[SPEC · seeded] … [→ <new>]`, records `state[tasks][new].from_delta=<prior>` lineage. (2) `drop-delta <task>` DISMISSES the first open SPEC delta → `[SPEC · dropped]`. NEW pure _resolve_spec_delta(text, new_status, pointer=None)->str|None [flips ONLY the first open line's status token via _SPEC_OPEN_TOKEN_RE + optional ` [→ ptr]` stamp; entry text + `(evidence: …)` byte-preserved; None when no open delta → caller refuses, no write] + _first_open_spec_text [first open delta's text, evidence stripped, from the SAME in-memory text the flip uses] + cmd_drop_delta + the `drop-delta` subparser + `--from-delta` on new-task/cmd_new_task. DESIGN-FOR-FAILURE: validate-ALL-then-write — resolve prior · prove ≥1 open SPEC delta · compute the seeded flip BEFORE any write; per-file _atomic_write; ordered new-TASK.md → prior-flip → state so the worst residue is 'new task created, prior still open & visible' [re-run blocked by the slug-exists guard → never a dangling [→] pointer at an absent task]. Additive verbs — no existing test enumerates the subcommand set, so no downstream test touched. §3 FROZEN @ v1 ([spec] first-open selector + [contract] no cross-file txn, both human-approved at the freeze). add.py logic+parser delta → ENGINE_MD5 7d908b2a → 5120b0c2, re-synced across all 3 copies; red suite test_seed_and_drop.py [8 tests], full suite green). Supersedes re-aimed @ spec-delta-grammar (delta-resolution 1/4: NEW SPEC-delta resolution track — a SEPARATE delta lifecycle from the five competency deltas. New grammar constants beside _DELTA_RE (single-source, no parallel grammar): _SPEC_STATUSES=(open,seeded,dropped) · _SPEC_DELTA_RE · _STATUS_SETS{tag→legal-status-set}. _spec_delta_entries + _collect_open_spec_deltas(root)->[{task,text,evidence}] (FLAT, open-only; SPEC resolves into a TASK, never consolidated into the foundation, so collected APART from _collect_open_deltas). _task_prose observe now derives first-open-SPEC → legacy 'Spec delta for the next loop:' field (back-compat) → '(unknown)'. _lint_task_deltas now lints BOTH '### Competency deltas' AND '### Spec delta', TAG-SCOPED via _STATUS_SETS: a cross-set status ([SPEC·folded], [SDD·seeded]) → unknown_status; evidence required on an OPEN entry of ANY tag; resolved-for-tag (folded|rejected · seeded|dropped) skipped (unknown tag defaults to the competency set so legacy folded lines still skip). cmd_deltas gained an 'open spec deltas' section + a separate --json 'spec'/'spec_total' key. TASK.md.tmpl §7 + _FALLBACK_TASK: the free-text 'Spec delta for the next loop:' line REPLACED by a '### Spec delta' block (no net new template comment — lean-pass <12 held). §3 FROZEN @ v1 (SPEC evidence REQUIRED, symmetric with competency — human-chosen at the freeze). add.py logic+parser delta → ENGINE_MD5 9c479010 → 7d908b2a, re-synced across all 3 copies; red suite test_spec_delta_grammar.py [11 tests]; downstream test_report observe tests + the template re-synced; full suite green. Supersedes re-aimed @ release-command (release-altitude 3/4: NEW GUARDED, record-only `add.py release <version>` — the ACT half of the 5th scope level (RELEASE, release.md), sibling to the release-report GATHER. cmd_release enforces the 4-code readiness floor THEN records the cut: release_security_open (≥1 open HARD-STOP gate — the UN-FORCEABLE reject, `--force` is ignored: a security finding is never shipped) · release_tests_red (an in-flight build w/o a recorded green gate — the tool-agnostic PROXY _build_in_flight(state): any active task phase∈{build,verify} AND gate==none, since the engine never shells to a test runner; `--force` overrides) · release_no_closed_milestone (_releasable==[]; `--force` overrides) · release_undisclosed_waiver (release_data['waivers'] non-empty AND no --with-waivers; `--force` overrides). VALIDATE-BEFORE-WRITE: all floor checks run before the first write so a reject leaves both files + state.json byte-for-byte unchanged. RECORD = exactly TWO writes via _atomic_write at the PROJECT ROOT (root.parent): CHANGELOG.md (prepend a `## <ver> — <date>` block, create w/ `# Changelog` header if absent) + RELEASES.md (prepend a newest-first row `## <ver> — <date>`·`milestones:`·`waivers:`·`evidence:`). The row's `milestones:` line IS the attribution (task-2 decision) — so cmd_release writes NO state.json (the `→ releasable` cue re-reads the ledger). Engine RECORDS, human SHIPS: NO git tag/publish/deploy, NO version-source bump. DESIGN-FOR-FAILURE: contents built in memory first; CHANGELOG written, then RELEASES.md (the commit point); a failed 2nd write ROLLS BACK CHANGELOG to its prior bytes (or unlinks it if created) and refuses release_write_failed — never a half-recorded cut. New helpers _build_in_flight · _prepend_block (newest-first, append-only — existing rows preserved verbatim) · _render_changelog_block · _render_releases_row; new `release` subparser (version positional + --force + --with-waivers + --evidence). test_min_pillar census: LIFECYCLE += `release 0.0.0` (refuses release_no_closed_milestone at its slot) + `release` ∈ _NONZERO_OK. §3 FROZEN @ v1 (risk:high, conservative gate). add.py logic+parser delta → ENGINE_MD5 d0d9b1ed → 9c479010, re-synced across all 3 copies; red suite test_release_command.py [12 tests], full suite green. Supersedes re-aimed @ release-report (release-altitude 2/4: NEW read-only `add.py release-report` — the GATHER half of the 5th scope level (RELEASE, release.md). release_data(root, state) clusters FIVE record-sets {releasable · changed · waivers · blockers · monitors + summary}: PURE·NO writes·exit 0 ALWAYS (only no_project non-zero)·text default + --json frozen facts interface — mirrors graduation_data/cmd_graduation_report. GATHER-not-JUDGE by construction: no readiness/score/ranking/verdict field (the FORBIDDEN-token guard). NEW helpers: _releases_path (the append-only ledger at the PROJECT ROOT = root.parent, a sibling of CHANGELOG.md — NOT inside .add/) · _released_milestones (the union of every `milestones:` row; FAIL-OPEN — a missing/unreadable/malformed ledger → empty set → every closed milestone reads still-releasable, so a vanished ledger never hides work) · _closed_milestones (LIVE status==done + ARCHIVED, all PASS-done) · _releasable (closed MINUS released — drives BOTH the status cue and the report) · _key_decisions_for (best-effort §Key-Decisions rows from PROJECT.md naming the slug, fail-open []). NEW module constants RELEASABLE_CUE + RELEASES_FILE. cmd_status gains the project-global read-only cue `→ releasable: N milestone(s) closed since last release` (fires for a live-done OR archived unreleased milestone — stands with no live milestones; additive; ledger read fail-open so a vanished ledger never silences it). NEW `release-report` subparser (--json/--plain) registered after graduation-report. Attribution source = RELEASES.md MEMBERSHIP (a milestone is released iff its slug is on a `milestones:` row) — chosen so the cue never reads compacted milestone files (the flagged riskiest decision). §3 FROZEN @ v1 (conservative gate). add.py logic+parser delta → ENGINE_MD5 91e76c35 → d0d9b1ed, re-synced across all 3 copies; red suite test_release_report.py [8 tests], full suite green. Supersedes re-aimed @ capture-evidence (udd-design-loop 3/4: NEW pure _missing_captures(root) -> list[str] [prototype names under .add/design/prototypes/ with no .add/design/captures/<name>.<ext> for ext in png·svg·jpg·jpeg·webp; PURE·TOTAL·READ-ONLY·doc-order; [] when proto_dir absent] + cmd_check appends a never-red WARN missing_capture per uncaptured prototype on the EXISTING `warnings` array — NEVER `checks`, so never feeds `failed`; silent-when-absent; the engine MEASURES capture presence, never renders. design.md beat 4 + "## Tool-agnostic capture" reopened ×3 skill trees + udd-wireframe.md "## Capture is evidence" ×2 template trees: captures at .add/design/captures/<name>.<ext> · attached/mentioned in the feature's TASK.md · @json-render/image [Satori→PNG/SVG] the named-default capture engine. Worked welcome/settings captures committed to .add/design/captures/ + cited in §6 as the demonstration [exit-criterion 4]. add.py logic delta → ENGINE_MD5 5599740d → 91e76c35, re-synced across all 3 copies; red suite test_capture_evidence.py [14 tests], full suite green. Supersedes re-aimed @ agent-update-nudge (1.4.0 release — integrate fork PR #17 [ewalliss]: the agent-agnostic stale-install nudge. ADD tells EVERY agent to run `add.py status`/`guide` first each session, so that is the one universal chokepoint to flag a stale install. main() gains a fail-open _maybe_nudge_update(args) preamble — STDERR-only, NEVER touches stdout or the exit code, gated to cmd_status/cmd_guide ONLY: it reads the launcher .add-version stamp [version+channel] and, when the registry latest > current, writes an ACTION-REQUIRED line naming the channel-correct command [npx @pilotspace/add@latest update · pipx run pilotspace-add update]. The engine's ONE deliberate network touch, bounded: 1.5s timeout, throttled once/24h via .update-cache.json, fail-open on offline/timeout/bad-payload [stale cache beats nothing], INERT without a launcher stamp [engine init writes no stamp, so the offline suite never touches the network], kill switch ADD_NO_UPDATE_CHECK=1. New helpers _read_json_safe · _write_json_safe · _version_gt · _fetch_latest_version · _cached_latest; imports += urllib.request, datetime += timedelta. IMPROVED over the raw PR: .update-cache.json added as a 3rd _GITIGNORE_BODY entry [transient-artifact hygiene] + new tests covering the cache-write & stale-cache-fallback branches. add.py delta → ENGINE_MD5 6c5ba081 → 5599740d, re-synced across all 3 copies; red suite test_update_nudge.py + extended test_add gitignore guards, full suite green. Supersedes re-aimed @ soul-artifact (v13-onboarding-polish 5/6: SOUL.md — a NEW survivor-layer living "voice" doc named Trusting — joins the foundation. add.py delta is TWO additive seams: SETUP_FILES += "SOUL.md" (last member) so cmd_init scaffolds .add/SOUL.md via the GENERIC render+write loop [never-clobber an existing one · skip-if-blank warns rather than write a 0-content survivor]; + a cmd_status VOICE pointer ["voice   : .add/SOUL.md (how I sound & what keeps your trust — read each session)"], existence-gated [no open/parse → no new IO failure path], sibling to the PROJECT.md context pointer. Ships templates/SOUL.md.tmpl [×3 byte-identical via test_bundle_parity + the .add dogfood copy] carrying the SOUL SCHEMA — Name · Tone · Communication style · Trust · Learns-from · Voice deltas — with a PROPOSED "Trusting" starter voice EXPLICITLY marked human-owned + overridable [the identity-values-human-owned guideline: tests assert the SCHEMA/mechanism ONLY, never the tone words, so the human rewrites the voice with zero build breakage] + a pointer to the soul-self-improve loop as the only voice-delta writer. + the skill SKILL.md "Always start here" gains the read-.add/SOUL.md-each-session instruction [skill ×3, test_tree_parity + test_bundle_parity]. test_add.test_init_creates_state_and_setup_files auto-covers SOUL.md [it iterates SETUP_FILES]. §3 FROZEN @ v1 [autonomous authorization; voice CONTENT surfaced OPEN to the human at the seam — soul-self-improve task 6 owns convergence to the human's wording]. add.py SETUP_FILES+cmd_status delta → ENGINE_MD5 eebbb443 → 6c5ba081, re-synced across all 3 copies; new red suite test_soul_artifact.py [8 tests], full suite green. Supersedes re-aimed @ dag-scheduler (v13-onboarding-polish 1/6: NEW read-only `add.py waves` DAG scheduler — _wave_schedule(state, mslug) [PURE·TOTAL: never mutates, never raises on dict input] + cmd_waves + the `waves` subparser. REUSES cmd_ready's satisfaction predicate [a dep is satisfied iff archived OR _task_done] and the _find_cycle idiom; groups the active milestone's not-done tasks into topological Kahn waves [wave = tasks whose open-member deps all landed earlier], names the CRITICAL PATH [longest member-dep chain, ties→smallest slug] via memoized depth, emits an advisory TIER HINT [top on the critical path, mid elsewhere — a proxy for scope-of-impact; run.md model pick stays advisory, never a gate], and surfaces a BLOCKED set [a not-done NON-member dep is unsatisfiable within this milestone → reported with waiting_on, never silently dropped]; an all-done/empty milestone schedules nothing [exit 0], the rejects are no_active_milestone · unknown_milestone · dependency_cycle [names the cycle members]. READ-ONLY: no save_state, no `next:` footer [read-only verb]. --json emits {milestone, waves, critical_path, critical_path_len, tiers, blocked}. streams.md gains the "## DAG strategy — let the engine schedule the waves" section [×3 byte-identical]. §3 FROZEN @ v1 [autonomous authorization]. add.py logic+parser delta → ENGINE_MD5 475d9d40 → 02426e9d → eebbb443 [v1.1: the VERIFY adversarial refute-read found a real HIGH bug — blocking did not propagate transitively, so a task whose only dep was a BLOCKED sibling was mis-placed in wave 1 as ready; fixed by computing the blocked partition to a FIXED POINT, +1 red guard test test_transitive_blocked_dep + 4 gap tests (empty milestone · self-dep cycle · linear-chain critical path · read-only across invocations)], re-synced across all 3 copies; red suite test_dag_scheduler.py [14 tests], full suite 1058 green. Supersedes re-aimed @ autonomy-command (NEW first-class `add.py autonomy show|set` verb — the autonomy level finally gains the CLI verb it lacked, so an agent under `auto`, applying the correct "first-class state has a command" model, can no longer hallucinate a phantom `add.py autonomy` and derail [the scla-mono symptom that opened this task]. cmd_autonomy + the `autonomy` subparser [positional action show|set · a1/a2 · --project · --yes; _opt_positionals=(a1,a2)]: `show` is read-only — prints declared · effective[fallback-resolved] · project default · the verify-gate owner [the SAME _driver_stop the footer/guide render]; `set` is the FIRST writer of the `autonomy:` header token — _autonomy_decl_line replaces the SINGLE declaration line idempotently [trailing comment preserved, NEVER appended — kills the scla-mono duplicate-line bug; inserts after slug:/heading when absent], _guard_autonomy_raise refuses a RAISE toward auto without --yes [autonomy_raise_unconfirmed — the engine's first confirm flag, the human-owned trust escalation], the risk:high guard [unguarded_high_risk_auto, reused _RISK_HIGH_RE + lowered-rung] runs BEFORE the write, a bad level → autonomy_level_invalid; reused VERBATIM _require_root + _resolve_task; state.json UNCHANGED [autonomy stays a header token]. De-command-shape: the 3 command-shaped "set autonomy: X" strings [garbled-project warning · the gate risk-guard · the check implicit_autonomy warn] + templates/TASK.md.tmpl now cite `add.py autonomy set`; a WORDING_RUBRIC.md idiom_map [enforced] fence bans the command-shaped idiom so it can never regress on the agent surface. §3 FROZEN @ v2 [v1→v2 change-request raised at TESTS: reject codes aligned to the guards it reuses]. add.py logic+parser delta → ENGINE_MD5 f15eb52a → 475d9d40, re-synced across all 3 copies; new red suite test_autonomy_command.py [16 tests] + sibling census ratifications [test_min_pillar LIFECYCLE += autonomy · test_rewrite_guides AUTONOMY_COMMAND_IDIOMS], existing engine suites stay green). Supersedes re-aimed @ gitignore-scaffold (vcs-hygiene: cmd_init now scaffolds a co-located .add/.gitignore — module constant _GITIGNORE_BODY ignoring scope-snapshot.json + pre-archive-state.bak.json (bare-filename patterns, any depth under .add/); written after the tasks/ mkdir, NEVER-CLOBBER an existing copy, additive — touches no snapshot/.bak on disk. add.py-only delta → 7b67afd1 → f15eb52a, re-synced across all 3 copies; guards: test_add.test_init_scaffolds_gitignore + test_gitignore_ignores_transient_artifacts (real git check-ignore enforcement) + test_init_never_clobbers_existing_gitignore. Supersedes re-aimed @ udd-check-lint (udd-design-foundation task 4/4: the COMPOSER — wire BOTH pure validators into `add.py check` + close the cross-file token resolution tasks 1+2 deferred. NEW cmd_check "UDD foundation" section (read-only, appended before the tally): discovers the named set under `.add/design/` (Fork A, human-chosen at the freeze over the `.add/` root) — root/design/tokens.json · root/design/catalog.json · root/design/prototypes/*.json; FAIL-CLOSED load (malformed JSON → malformed_tokens_json · malformed_catalog_json · malformed_prototype_json, never _die, never a crash); composes _token_layer_violations(tokens) + _catalog_tree_violations(catalog, tree) UNCHANGED, one FAIL per violation (named code as reason) / one PASS per clean file; SILENT-WHEN-ABSENT (no named set → zero new output, so clean/non-UI projects + this dogfood repo stay green). NEW pure _prop_token_resolution_violations(tokens, catalog, tree) (+ helper _semantic_token_index) closes the deferral: resolves every tree token-prop alias targeting `semantic` against tokens.json — terminal node absent / not a token → unresolved_prop_token; the resolved semantic token's INHERITED $type (DTCG: $type on the GROUP, leaf carries only $value) ≠ the catalog PropSpec's `token:<$type>` → prop_token_type_mismatch. PURE + TOTAL (never mutates, never raises on dict inputs); acts ONLY on a catalog token-PropSpec × a `{semantic.*}` tree alias (the props task 2 passed LAYER-only) — non-token/non-alias/non-semantic/uncataloged are SKIPPED (task 1/2 own those codes; no double-flag). cmd_check's read-only-ness + output format (PASS/FAIL/WARN · "check: N passed, M failed (K warnings)" · the --json schema) + SystemExit(1)-iff-failed are PRESERVED — the ~30 existing cmd_check tests stay green. Fork A reconciliation (human's freeze choice): the DESIGN.md.tmpl + udd-catalog.md live-file pointers move to `design/…` (×3 each; the task-3 substring tests survive — `design/tokens.json` still contains `tokens.json`). new red suite test_udd_check_lint.py (18): the resolver unit tests (clean · the shipped samples cross-resolve clean · unresolved · $type-mismatch · group-alias · skip-non-token/non-semantic · skip-malformed-PropSpec · skip-malformed-resolved-$type · purity) + the cmd_check integration over .add/design/ (clean-passes-read-only · each named red · fail-closed malformed tokens/catalog/prototype · silent-when-absent). add.py-only logic change → ENGINE_MD5 e2e1dd0 → 7e0ad6a → 7b67afd1 (v2 — the verify-gate adversarial refute found the cross-file resolver DOUBLE-FLAGGED two shapes the §3 'no double-flag — task 1/2 already own those codes' clause forbids: a malformed token PropSpec [bad / absent `token` → task-2's malformed_catalog] and a resolved semantic token whose inherited $type is malformed / absent [→ task-1's unknown_type]. Fix: the resolver now SKIPS a prop whose declared `token` ∉ _TOKEN_TYPES AND skips the mismatch arm when the resolved token's $type ∉ _TOKEN_TYPES — both routed to their task-1/2 owners; the unresolved-alias + legit-mismatch paths are UNCHANGED. +2 red no-double-flag resolver tests + 2 fail-closed integration tests for malformed_catalog_json · malformed_prototype_json [§1 Reject codes that had shipped untested] + the clean test strengthened to assert the prototype PASS line; 14→18 tests). This pin supersedes re-aimed @ udd-design-template (udd-design-foundation task 3/4: the prose FRONT-DOOR — wire a DESIGN.md template into setup. ENGINE delta is ONLY SETUP_FILES += "DESIGN.md" (appended last) so cmd_init renders the NEW templates/DESIGN.md.tmpl via the GENERIC L377–389 render+write loop → .add/DESIGN.md (skip-not-clobber; a blank render is warn-skipped). The template is a STRUCTURED BINDING doc (Fork A) — SIX required sections: a header + a "no UI? this doc is optional — delete it" self-note · ## Identity (brand·palette·typeface·voice as HTML-comment PROMPTS, NO pre-filled brand value — the identity-values-human-owned guideline; a concrete hex outside a prompt = identity_prefilled) · ## Principles/persona · ## Screens (a prompted prototypes/<name>.json index SEEDED with prototype.sample.json — the human-added 6th section at the v1 freeze, "Freeze + Screens section") · ## Foundation (pointers binding the named set tokens.json/catalog.json/prototypes/ + their docs udd-tokens.md/udd-catalog.md + the shipped samples tokens.sample.json/catalog.sample.json/prototype.sample.json, Fork B — start from the samples) · ## Render (a pointer to udd-catalog.md's "## Render recipe"). Rendered with {{project}}/{{stage}}/{{date}}; MUST render non-blank. FIVE named reds (the DESIGN.md guards; udd-check-lint does NOT lint this prose): design_not_wired · design_template_blank · identity_prefilled · design_section_missing · mirror_drift. Fork C: wiring is UNCONDITIONAL — every init drafts DESIGN.md, and the self-note tells a non-UI project to delete it; the engine delta is ONLY the one SETUP_FILES tuple member (the render+write loop is untouched). Ships templates/DESIGN.md.tmpl (canonical+bundle byte-identical via test_bundle_parity, + the .add dogfood copy for the dogfood engine) + the phases/0-setup.md "Draft to the lock" DESIGN.md line (skill ×3, test_tree_parity + test_bundle_parity) + test_init_creates_state_and_setup_files auto-covers DESIGN.md (it iterates SETUP_FILES). add.py-only one-tuple-member delta → ENGINE_MD5 3cdfaab → e2e1dd0. No contract change-request — the §3 freeze added Screens at v1 before the stamp. This pin supersedes re-aimed @ udd-catalog-content-schema (udd-design-foundation task 2/4: the render-ready half — a NEW pure validator _catalog_tree_violations(catalog, tree) -> list[(code, path, detail)] (+ helpers _propspec_malformed, _prop_value_code) checking OUR compact-JSON component CATALOG against a flat json-render content TREE. CATALOG = {components: {<Name>: {description?, hasChildren?(default false), props: {<prop>: PropSpec}}}}; a PropSpec is exactly one of {type: string|number|boolean} · {type: enum, values:[str]} · {type: token, token:<$type>} ($type ∈ the task-1 set color·dimension·number·fontFamily·fontWeight·duration). CONTENT TREE = json-render flat Spec pinned @ vercel-labs/json-render v0.19.0 / commit 4e4dc46 (the milestone's named top risk: young-project schema drift): {root, elements: {<id>: {type ∈ catalog · props ⊆ declared · children? only if hasChildren, ids ∈ elements}}}; a token-prop value is the {semantic.dotted.path} alias; json-render's state/on/visible/repeat are PASSED THROUGH (render-compatible + clickable) but NOT linted, keeping the validator lean. EIGHT named reds: tree_cites_uncataloged_component · unknown_prop · prop_type_mismatch (incl. a token prop given a non-alias literal) · non_semantic_prop_token · dangling_child · children_not_allowed · missing_root · malformed_catalog. PURE (never mutates inputs), stdlib-only, DETERMINISTIC document order; [] == valid. CHANGE-REQUEST raised at TESTS → frozen §3 @ v2 (Option B, human-approved): the frozen (catalog, tree) signature cannot resolve a token's existence/$type (that needs tokens.json), so non_semantic_prop_token was NARROWED to LAYER-only — a token-prop alias MUST target the `semantic` layer (first path segment); target existence + $type-match DEFER to udd-check-lint, the composer that holds tokens+catalog+tree. A strict narrowing; the signature + the other 7 codes are UNCHANGED, no behavior added; §1 reject + §2 scenario amended to match. Fork A/D (human-approved at the freeze): the validator is NOT wired into cmd_check here — udd-check-lint LATER calls BOTH _catalog_tree_violations + _token_layer_violations inside cmd_check (named reds) plus the cross-file token resolution; cmd_check is NAMED-not-touched. Foundation shape = a NAMED SET — tokens.json · catalog.json · prototypes/<name>.json (Fork A). Ships templates/catalog.sample.json + templates/prototype.sample.json (a clean worked pair: Screen > Card > {Text, Text, Button}, token props bound to semantic tokens, a Button on-click passthrough) + templates/udd-catalog.md (the render recipe: a ~20-line catalog.json → json-render defineCatalog adapter; the flat tree feeds catalog.validate AS-IS), all ×3 byte-identical (test_bundle_parity + test_tree_parity). add.py-only logic change → ENGINE_MD5 c107344 → 774e025 → 3cdfaab (v3 — adversarial refute at VERIFY found a HARD-STOP crash + a structural fail-open, Option A close: the validator is now a pure TOTAL function that NEVER raises; a non-object component entry → malformed_catalog; a 9th code `malformed_element` flags a tree element / props / children of the wrong JSON type; an empty `children` array == absent. Strict fail-closed additions — the signature + the prior 8 codes' behavior are UNCHANGED; §1 reject + §2 scenarios amended; +3 red tests + the 4 refute-flagged weak tests strengthened). This pin supersedes re-aimed @ udd-token-schema (udd-design-foundation task 1/4: NEW pure validator _token_layer_violations(tokens) -> list[(code, path, detail)] + helper _token_value_form_ok — the compact-DTCG 3-layer token dialect's shape/citation checker. SIX named reds: unknown_layer · unknown_type · unresolved_alias · cross_layer_citation · primitive_has_alias · malformed_value. A token's LAYER is its top-level group name (primitive/semantic/component); fail-closed citation rule (primitive = literal only · semantic → primitive only · component → semantic only; aliases are the DTCG `{layer.dotted.path}` curly form, chains allowed, must terminate in a literal). Compact SCALAR value forms (color "#RRGGBB"/"#RRGGBBAA" · dimension "<n><unit>" · number · fontWeight 100-900|keyword · duration "<n>ms|s" · fontFamily str|[str]) DIVERGE from DTCG 2025.10's value objects — every divergence NAMED in templates/udd-tokens.md, which cites the Format Module 2025.10 Final CG Report. PURE (never mutates input), stdlib-only, DETERMINISTIC document order; [] == valid. Fork A (human-approved at the §3 freeze): the validator is NOT wired into cmd_check here — udd-check-lint LATER calls it inside cmd_check as named reds + adds the catalog/tree rules; cmd_check is NAMED-not-touched. Ships templates/tokens.sample.json (a clean 3-layer worked example) + templates/udd-tokens.md, both ×3 byte-identical (test_bundle_parity + test_tree_parity). add.py-only logic change → ENGINE_MD5 a4dcc0b → c107344 (8329cd4 was the pre-refute build; the verify-gate adversarial refute found a fail-OPEN gap — a token node carrying non-$ children skipped its subtree — so _index/_walk now ALWAYS descend, validating nested children fail-closed; guarded by test_nested_token_children_are_validated). This pin supersedes re-aimed @ gate-owner-marker (next-step-seams task 2/3: the footer + the `guide` TEXT next-step line now NAME the driver — ` [you drive]` (the AI proceeds) / ` [human gate]` (a human owns it) — filling the slot next-footer-engine reserved. ONE resolver, no second owner map: NEW _driver_stop(root, state, slug, phase) = _phase_owner(phase) != "ai" for EVERY phase EXCEPT verify, where it is _effective_autonomy(root, state, slug) != "auto" (the single dial seam — auto auto-gates → the AI; conservative/manual → the human; run.md:21 the contract freeze stays human dial-blind). _driver_marker(stop) renders the reserved word. _effective_autonomy reads the task's own `autonomy:` line, falling back to _project_autonomy when unset(None)/unknown("?") — never silently auto. _next_footer fills the slot: Arm A via _driver_stop(phase×autonomy); Arm B via the NEW _decide_next_pair(state, d) -> (text, human_stop) [the precedence extracted ONCE; _decide_next_base is now a thin [0] wrapper so the report rollup/digest callers are unchanged] — every decision point (decompose · HARD-STOP · goal-not-met · consolidate/archive · approve-contract · gate) → [human gate], the run-in-progress fallthrough → [you drive]; fail-soft carries NO marker (never assert a driver it could not compute). cmd_guide TEXT prints the marker on its `next   :` line (computed AFTER the unknown-phase guard, so a bad phase fails clean — the marker invents no default). Option F (human-confirmed at the freeze): the FROZEN machine-state-json JSON `owner`/`stop` is UNTOUCHED — changing it is a change-request against THAT contract (intake.md); the marker rides the footer + guide-TEXT only, and the guide-text↔guide-JSON divergence at verify-auto is a recorded §7 delta, not absorbed here. Sibling co-updates in test_next_footer_engine.py (§5 scope, both strengthenings): Arm B HARD-STOP exact-equal now pins ` [human gate]`; test_marker_slot_reserved→test_marker_slot_filled now asserts ` [you drive]`. This pin supersedes re-aimed @ next-footer-engine (whose ONE-engine-sourced-footer mechanics it preserves verbatim, described next), which itself carried — Arm A names the active in-flight task's phase command (advance, or the gate verbs at verify, guarded gate=="none" AND phase!="done" so a HARD-STOPped task falls to Arm B not back to re-gate) + its PHASE_GUIDE why; Arm B reuses _decide_next_base over the active-milestone rollup — the SAME precedence the report renders — with the empty-rows branch reshaped command-first ("decompose into tasks — add.py new-task <ms>", replacing the dead-end "none — no tasks yet"). Wired into advance·gate·new-task·new-milestone·phase·use·reopen·set-milestone·milestone-done·lock·stage·archive-milestone·compact, converging their ad-hoc tails (new-task "then: advance", new-milestone "Decompose…", gate HARD-STOP "return to BUILD", milestone-done "Confirm…") so each prints exactly ONE next-step. EXCEPTIONS (human-confirmed at tests, both keep their existing seam, no footer): the exit-3 HEAL paths (cmd_heal · gate→heal raise SystemExit(3) with a return_to_build redo message) and cmd_init (its bootstrap "next: run /add"/"lock" block already names the setup-flow next step; the generic resolver would regress it). Fail-soft: the footer is computed AFTER save_state and never crashes a saved mutation — no active milestone / unreadable doc / corrupt rollup degrade to "next: add.py status — re-orient". Pure render, writes nothing; the trailing marker slot is reserved for the sibling gate-owner-marker. read-only surfaces (status/guide/report) unchanged. _scope_guard no longer dies in place on the RECOVERABLE findings — an out-of-scope touch (source 'scope') and a present-but-wrong sidecar (diverged|unparseable, source 'scope-tamper') call the SAME _heal_or_escalate the tamper tripwire uses, returning the task to BUILD for an honest redo (exit 3) and counting against the ONE shared per-task HEAL_CAP; the (CAP+1)th confirmed finding records gate=HARD-STOP and escalates heal_exhausted. The ERASED baselines stay die-in-place (exit 1, no heal): a MISSING sidecar is scope_snapshot_tampered and an erased anchor under a present sidecar is scope_anchor_missing — a redo cannot recreate erased evidence (tripwire_missing parity). Every heal reason CARRIES its named code so the refusal-token assertions still match; placement unchanged — after _tamper_guard, BEFORE the waiver write, off the HARD-STOP path; a heal never rewrites the sidecar or anchor. check stays a read-only WARN monitor — never routes to heal, never increments. This pin supersedes re-aimed @ scope-violation-heal (whose recoverable-heal mechanics it preserves verbatim, described above). Carries the prior scope-gate-enforce gate (NEW _declared_scope parses the frozen §5 grammar — backticked tokens · FIRST 'Scope (may touch):' line · ./… task dir · '/' project root · bare sibling-of-previous · outside-root dropped · directory token = WHOLE subtree · no line = UNDECLARED grandfathered · garbage = [] no cover; tests->build _scope_walk {rel: md5} pruning _SCOPE_EXCLUDE_DIRS into sidecar scope-snapshot.json anchored by sidecar-md5; v2 anchor×sidecar co-witness after the anchor-erase REFUTE; v3 UNDECLARED-crossing cleanup), the engine-merge-base-enforcement wave-ledger gate (cmd_check wave section + read-only cmd_wave_verify, v4 status-field grammar, exactly-one echo column, six drift fixtures), and the engine-argv-portability re-bind. tool-agnostic — never runs git)) · re-aimed @ fold-suggestion-seams (surface the fold-into-PROJECT.md suggestion live in the loop: cmd_advance prints a per-task `add.py fold --task <slug>` hint at the OBSERVE transition where the loop lessons are captured; cmd_milestone_done previews the open lessons by competency + emits the ready-to-run fold command. Additive prose only — no decision path changes; the verb word is interpolated via _FOLD_VERB so the domain wording-lint sees no bare slang. ENGINE_MD5 fa8e9818 → 1911b250)
-ENGINE_PKG_MD5 = "7ec748db0625ad4951092af9b972af26"  # re-aimed @ extract-io-state (io_state.py joined the package). engine-package-skeleton: manifest digest over add_engine/*.py (engine_manifest.package_digest); pins the package modules so ENGINE_MD5 can stay md5(add.py) (the ~52 EnginePinTest copies rely on that). A LITERAL — engine_pin never hashes.
+ENGINE_MD5 = "6d731aabd945f85921007d39cea532d0"  # re-aimed @ extract-io-state (engine-modularization 2/N: 4 IO primitives moved to add_engine/io_state.py, add.py shrank). ENGINE_MD5 is md5(add.py) — re-aimed @ engine-package-skeleton (engine-modularization 1/N: ENGINE_MD5 is now a MANIFEST DIGEST over the package — add.py + add_engine/*.py — computed by the separate engine_manifest.engine_digest(); the 43 public + 6 private constants moved verbatim to add_engine/constants.py, add.py re-exports them; pin stays a literal). prior: re-aimed @ phase-build-guard (audit-hardening, F4: `add.py phase build` is no longer a backdoor around the tests->build gate stack — the `nxt=="build"` guards+snapshots are extracted into a shared `_build_entry(root, state, slug)` helper that BOTH cmd_advance and cmd_phase call, so the admin override runs the IDENTICAL freeze gate · build-expectations · unflagged-freeze + flag stamp · tamper tripwire · §5 scope snapshot; verify's _tamper_guard is now armed on a cmd_phase-entered build; cmd_advance behavior byte-unchanged [pure extraction]; heal loop sets phase=build directly so stays exempt). prior: setup-commit-prompt (cmd_init closes with a one-line tip — both greenfield + brownfield branches — reminding the user to commit the .add/ folder to git so the team shares the ADD state; additive output only, no behavior/gate change). prior: gitignore-bak-seed (.add/.gitignore now lists `pre-update-state.bak.json` — the installer update backup — single-sourced in tooling/templates/gitignore.tmpl kept byte-identical to _GITIGNORE_BODY [parity test]; both installers seed/refresh .add/.gitignore on install+update via _seed_gitignore [pip] / seedGitignore [npm]: seed-if-missing ELSE append-if-absent the missing engine-transient pattern lines, additive-only + idempotent + fail-soft). prior: loose-task-release (release-attribute milestone-free DONE tasks: a "loose task" = milestone-free + done [_task_done] + not in any RELEASES.md `loose tasks:` row; NEW _released_loose_tasks / _releasable_loose_tasks readers [parallel to the milestone track, fail-OPEN]; release_data gains a "loose" record-set + summary["loose"]; _render_releases_row writes a `loose tasks: <slugs|none>` line [kwarg, default none]; cmd_release no-op guard now fires only if NO milestones AND NO loose tasks [a loose-only cut works]; status prints a SEPARATE loose cue; release_security_open floor UNCHANGED). prior: todo-capture (new `add.py todo` backlog verb: `todo "<text>"` captures a jotted idea into state["todos"] [{id=max+1, text, created, status}], bare `todo` LISTS open items, `todo --done <id>` closes one; todo_empty / todo_unknown guards; a todo is NOT a task [no spec/contract/gate] — promote via new-task --fast. additive state key, drift-tolerant). prior: standalone-fast-task (blessed milestone-free fast lane: cmd_new_task AFFIRMS a --fast task created with no owning milestone — "standalone fast task ... milestone-free by design" — instead of the intake nudge; cmd_check gains an ADDITIVE `infos`/`informed` tier so a fast standalone is a soft INFO, NEVER a WARN/orphan; the non-fast orphan path is UNCHANGED [test_v8_1_orphan_guard green]). prior: multirepo-federation (`federate pull <id>` lands a producer repo's published immutable contract snapshot at the local .add/contracts/<id>.json [the cross-REPO transport]; `_federation` reader; fail-loud unknown/missing/invalid/version-mismatch; cmd_check federation_source_unreadable WARN; opt-in byte-identical). prior: cross-component-milestone (cmd_advance scenarios->contract HOLD: a `consumes:` task of a DECLARED contract is refused producer_contract_unfrozen until the producer snapshot exists [the intra-milestone BE->FE ordering]; opt-in byte-identical). prior: cross-component-contract+refute-heal (null-hash snapshot now HARD-STOPs (fail-loud Finding 2); a corrupt LIVE snapshot now surfaces contract_snapshot_unreadable WARN not silent no-stale (Finding 3). NEW _contracts/_task_produces/_task_consumes/_contract_snapshot/_contract_body_hash/_contract_findings + produces:/consumes: header regexes; cmd_advance contract->tests crossing WRITES .add/contracts/<id>.json on a `produces:` freeze [idempotent, atomic] + PINS the live hash on a `consumes:` cross [missing snapshot HARD-STOPs contract_snapshot_missing]; cmd_check adds contract_producer_unknown RED + contract_consumer_stale WARN. Hash is over the §3 fenced SHAPE so a pure version bump does not churn. Opt-in byte-identical. prior: per-component-verify v3 (NEW pure _cite_region: the Build-expectations evidence region [matches the standard ### heading AND the fast-lane bare line, up to GATE RECORD] with the engine stamp STRIPPED — closes refute-read #3 Finding 1 [stamp landing inside the block self-satisfies] + Finding 2 [fast-lane template has no ### heading -> component-bound fast task permanently false-blocked]; cmd_gate cite-check now calls _cite_region; prior v2 (cite-search SCOPED to the §6 "### Build expectations" block — subsumes the v1 stamp-strip AND closes refute-read Finding 2 [a generic green_bar like "all tests pass" that collides with §6 checklist/Outcome boilerplate no longer self-passes]; prior: refute-heal (cite-check now STRIPS the engine's own "·expected green-bar:" stamp before the §6 substring search, so a prior HARD-STOP's stamp can't self-satisfy a later PASS [BLOCKER from refute-read]; component-aware-add M2 per-component-verify: NEW pure _task_green_bar; cmd_gate gains a completing-only cite-gate (component_green_bar_uncited) for a bound task with a green_bar — surfaces + stamps "component: X · expected green-bar: B" into §6 via _stamp_gate_record; cmd_check WARNs component_green_bar_unset. Engine still never runs a suite. Unbound/no-green_bar = byte-identical (every branch guarded behind _task_green_bar is not None; HARD-STOP exempt). risk:high. red suite test_per_component_verify.py [9]). PRIOR ⤵ re-aimed @ component-registry [self-heal: degrade-safe tasks/ scan in _component_findings + reserve "?" component name — refute-read MAJOR+MINOR] (component-aware-add M1 component-registry: NEW component registry — readers _components/_component_root/_task_component/_component_findings + _COMPONENT_LINE_RE; _declared_scope EXTENDED additively to append a bound component's root (compose, never redraw explicit-token resolution); cmd_check gains fail-closed RED registry findings (components_malformed/component_unknown/component_root_outside, like wave_ledger_malformed); TASK.md.tmpl gains a commented `component:` affordance; guarded `import tomllib`. OPT-IN — no .add/components.toml ⇒ byte-identical to pre-component ADD. risk:high. red suite test_component_registry.py [18]). PRIOR ⤵ re-aimed @ loop-aware-orient (status+guide STEER into the loop at the loop juncture. NEW pure/fail-soft fn _done_resume(root,state,slug)->(headline,next_step,chapter): classifies a DONE task from its milestone exit-criteria tally — LOOP-JUNCTURE total>0&met<total -> name the unmet goal + route 09-the-loop.md; GOAL-MET total>0&met==total -> milestone-done; PLAIN (no criteria/no milestone/read error) -> today's "start the next feature"/02-the-flow.md, BYTE-IDENTICAL. cmd_status done-branch + cmd_guide human AND --json done-branch route via the helper, overriding only when chapter!=02-the-flow.md so PLAIN output is unchanged. loop.md cue HEALED: the engine now prints "goal not met (m/n exit criteria)" so the doc's status-attribution is TRUE (no doc edit needed). red suite test_loop_aware_orient.py [9]. risk:low — additive + fail-soft). PRIOR ⤵ re-aimed @ ccsk-rule-file (re-integration of fork PR #45 onto post-1.9.0 main: ccsk-aware rule-file placement. NEW module consts RULES_FILE_REL=.claude/rules/add-workflows.md + WORKFLOW_HEADINGS + _RULE_REF_LINE; NEW pure/fail-soft fns _rule_file_mode (disk-derived: --rule-file OR a .ccsk/ dir OR an existing rule file — no persisted state) + _strip_inline_block (migrate an inline ADD:BEGIN..END out; an unterminated BEGIN is left as-is) + _insert_rule_reference (bullet under an existing Workflows/Rules heading at any # level, else append a fresh ## Workflows) + _ensure_claude_reference (strip inline + add the ref; .bak on change; idempotent). _inject_guidelines gains a rule_file param + a CLAUDE-only branch: the full block goes to .claude/rules/add-workflows.md and CLAUDE.md keeps a reference bullet — AGENTS.md stays inline. cmd_init + cmd_sync_guidelines pass getattr(args,'rule_file',False); a --rule-file store_true is added to the init + sync-guidelines parsers. WITHOUT the flag / a .ccsk dir / an existing rule file: output+scaffold BYTE-IDENTICAL to today. red suite test_rule_file_mode.py [11] + test_installer_prompts.py [+4 pip/npm parity]. risk:low — additive + fail-soft. credit: original feature by Crystal D. (@imnortheastt)). PRIOR ⤵ re-aimed @ fast-new-task-flag (fast-lane 3/4: `add.py new-task <slug> --fast` opts into the fast lane. cmd_new_task renders the MINIMAL `_render_template("TASK.fast.md", …)` (vs "TASK.md") + records the durable marker `state["tasks"][slug]["fast"] = True` (ABSENT on a normal task — grandfathered). The `new-task` parser gains a `--fast` store_true. cmd_status tags an ACTIVE fast task with a ` · fast` marker on the `active  :` line (presentation-only, existence-gated). v1→v2 CHANGE-REQUEST (human-chosen "fast implies floor" at the re-freeze): v1's §3 wrongly claimed the freeze floor was "enforced upstream, UNCONDITIONAL, no new guard" — FALSE, since the existing `unflagged_freeze` check only fires once §3 is ALREADY frozen, so a DRAFT-§3 task reached gate=PASS unrefused. v2 WIRES `--fast` into the freeze-before-build gate (shipped by task 2): the gate's freeze-firing condition becomes `_freeze_gated := _optin OR state["tasks"][slug].get("fast") is True`, so a fast task CANNOT cross tests→build with a DRAFT §3 under ANY milestone (opted-in or not) — collapse-never-skip is REAL for the fast lane itself. The build-expectations gate stays opt-in-only (`_optin`). `--fast` composes with `--from-delta` (both pre-fill §1 Feature; the fast template has it). WITHOUT `--fast`: output + scaffold + state BYTE-IDENTICAL to today (renders TASK.md, no `fast` key) — zero ripple, no existing task carries the marker. risk:high · §3 FROZEN @ v2 (approved by Tin Dang). red suite test_fast_new_task_flag.py [8 tests: minimal-scaffold+marker · plain-new-task-unchanged · fast floor holds under a PLAIN milestone (contract_not_frozen) · plain task NOT freeze-gated (zero ripple) · fast completes through gates to PASS+§6 stamped · status marks/omits the fast tag · --fast ∘ --from-delta]. add.py logic+parser delta → ENGINE_MD5 80985fc3 → d4807ff9, re-synced across all 3 copies; full suite green. PRIOR re-aim @ freeze-before-build-gate (fast-lane 2/4: makes "collapse-never-skip" REAL — a NEW opt-in freeze-before-build gate in cmd_advance at the `nxt == "build"` crossing, placed BEFORE the existing build-expectations gate. GUARD: `_optin := task has a milestone M AND state["milestones"][M]["await_confirm"] is True`; `if _optin and not _contract_frozen(_raw_phase_bodies(root, slug).get(3, "")): _die("contract_not_frozen")`. Mirrors the build-expectations gate's opt-in switch EXACTLY (the same --await-confirm marker, one level out) — a plain / no-await_confirm / milestone-less task is NEVER gated, so every existing advance-to-build flow stays byte-green (ZERO ripple, MEASURED). Refactor: `_ms`/`_optin`/`raw3` computed ONCE at the build crossing and reused by both opted-in gates (freeze, then build-expectations) + the existing unflagged_freeze flag check. validate-then-write — the refusal runs BEFORE the tripwire/scope snapshots, writes nothing, task stays at `tests`. EMPIRICAL JUSTIFICATION: the engine did NOT hard-enforce freeze-before-gate for ANY task before this — a DRAFT §3 task reached gate=PASS unrefused (the freeze was a skill/human convention only); an ALL-tasks guard breaks ~40 fixture files (81 fails at build-entry / 176 at gate — the confirm-parent trap), so the human chose the opt-in path. SIBLING-FIXTURE FIX (the ONLY measured ripple): test_build_expectations_gate's `_optedin_task_at_tests` helper now freezes §3 (+ a well-formed least-sure flag) so its 2 opted-in fixtures clear the new gate and still reach the build-expectations check. red suite test_freeze_before_build_gate.py [6 tests: opted-in unfrozen→contract_not_frozen + stays tests + no scope-snapshot · opted-in frozen+filled→advances · plain milestone unfrozen→advances · no-milestone unfrozen→advances · freeze PRECEDES build-expectations (DRAFT §3 + placeholder §6 → contract_not_frozen not build_expectations_unfilled) · frozen §3 + placeholder §6 → build_expectations_unfilled takes over]. risk:high · §3 FROZEN @ v1 (approved by Tin Dang). add.py logic delta → ENGINE_MD5 280daae8 → 80985fc3, re-synced across all 3 copies; full suite green. PRIOR re-aim @ fast-lane-template (fast-lane 1/3: NEW `_FALLBACK_TASK_FAST` minimal-template circuit-breaker analog [sections {0,1,3,4,5,6}; §2 SCENARIOS + §7 OBSERVE dropped] + `_render_template` fallback branch extended to "TASK.fast.md" via a name→fallback map [TASK.md + TASK.fast.md], so a deleted templates/ never hard-fails the fast lane. Ships templates/TASK.fast.md.tmpl ×3 byte-identical [canonical + dogfood + _bundled]; the full TASK.md.tmpl is byte-untouched; the --fast flag + cmd_new_task wiring + check/audit tolerance are the NEXT task [fast-new-task-flag]. The minimal template keeps the trust floor: §3 freeze-flag + Status [_flag_well_formed], §6 GATE RECORD/Outcome/Reviewed [_stamp_gate_record], §0 Anchors [_grounded_state], §4 red-before-build, §5 Scope. risk:high · conservative gate · §3 FROZEN @ v1 (approved by Tin Dang). red suite test_fast_lane_template.py [25 tests]. add.py delta → ENGINE_MD5 cb7ddd03 → 280daae8, re-synced across all 3 copies; full suite green. PRIOR re-aim @ gate-record-writeback (flow-enforcement 3/3 — FINAL: the gate write-back — `add.py gate <outcome>` now mirrors the resolved verdict into the task's §6 `### GATE RECORD`, so the file and state.json never silently diverge (Finding C: `gate PASS` recorded state but left the §6 a `<…>` placeholder, which `audit` flags `malformed_gate_record`). NEW `_stamp_gate_record(root, state, slug, outcome)` called in cmd_gate AFTER save_state (state is the source of truth; the file only mirrors it): rewrites the §6 `Outcome:`/`Reviewed by:` lines (+ the `If RISK-ACCEPTED -> owner·ticket·expires` line from the state waiver) to resolved values via `_actor_stamp(state).name` + `date.today()`. NO opt-in — UNLIKE the two refusal gates (tasks 1&2), a write is ADDITIVE and never refuses, so it cannot break the census; the SAFETY is GRANDFATHER: each line is rewritten ONLY while it still holds a `<…>` placeholder, so a hand-filled/already-resolved record is byte-untouched, and a re-gate never clobbers. No GATE RECORD block / no placeholder / unreadable file → silent no-op, the file stays byte-identical (a completing gate still syncs the `phase:` marker as before — orthogonal). The audit divergence detector (`malformed_gate_record` ≠1 Outcome line · `gate_record_mismatch` §6≠state) was ALREADY present — the write-back is what finally makes opted-in-or-not tasks PASS it, project-wide. §3 FROZEN @ v1 (all-tasks, no opt-in — human-chosen at the freeze over the opt-in draft). risk:high · conservative gate. red suite test_gate_record_writeback.py [6 tests]. Build correction (faithfulness, surfaced at build): the "no §6 block → file byte-unchanged" scenario over-claimed — a completing `gate` ALWAYS rewrites the `phase:` marker (verify→done) independent of write-back; the test normalizes that one line and additionally asserts no GATE RECORD block was fabricated. Prose lint: the `_stamp_gate_record` docstring's "two refusal seams" tripped the ubiquitous-language guard (seam = a Group-C machine constant, banned as prose) → reworded "gates". add.py logic delta → ENGINE_MD5 d7a104fa → cb7ddd03, re-synced across all 3 copies; full suite green. PRIOR re-aim @ build-expectations-gate (flow-enforcement 2/3: the tests->build verify-expectations gate — `cmd_advance` (build-entry) now refuses `build_expectations_unfilled` when an OPTED-IN task's §6 `### Build expectations` block is a `<…>` placeholder/empty, BEFORE the tripwire/scope snapshots (validate-then-write — a refused advance leaves phase=tests, no sidecar). OPT-IN = the task's PARENT milestone has a `confirmed` key (the SAME --await-confirm switch task-1 uses, one level out) — so `--await-confirm` now arms the WHOLE flow-enforcement suite (content-aware confirm + build-expectations) per milestone, and the 33 existing advance-to-build suites (plain/no milestone) stay green = ZERO ripple. `_section_unfilled` EXTENDED compatibly: break the section at ANY next header (`#`, not just `## `) + SKIP `>` blockquote guidance lines, so it reads the `### Build expectations` SUB-section; task-1's `## Shared / risky contracts` truth table is preserved (test_contract_fill_gate stays green). risk:high · conservative gate · §3 FROZEN @ v1 (milestone-inherited opt-in, approved by Tin Dang). red suite test_build_expectations_gate.py [6 tests]. OPT-IN MARKER (v1.1, caught by the full suite at build): keying the gate on `"confirmed" in m` broke the census — `milestone-confirm` stamps confirmed:true on a PLAIN milestone too, so at advance time a never-opted-in milestone looked opted-in. Fix: `new-milestone --await-confirm` now ALSO seeds a stable `await_confirm:true` marker (set ONLY at creation, never by milestone-confirm), and the build gate keys on `milestone.await_confirm is True`. add.py logic delta → ENGINE_MD5 428ca1d1 → ee43743c → d7a104fa, re-synced across all 3 copies; full suite green. PRIOR re-aim @ contract-fill-gate (flow-enforcement 1/3: the CONTENT-aware confirm gate — `milestone-confirm` now refuses `milestone_contracts_unfilled` when an OPTED-IN milestone's MILESTONE.md `## Shared / risky contracts` section is still a `<…>` placeholder/empty, so "confirmed" MEANS the cross-task contracts were present. NEW pure `_section_unfilled(md_text, header)` [PRESENT-but-(empty|`<…>`)→True; ABSENT→False grandfather; filled→False] — the shared placeholder predicate the sibling `build-expectations-gate` reuses for §6. Gate is OPTED-IN ONLY (`"confirmed" in m`): a grandfathered no-key milestone keeps the plain stamp, so the census + every existing flow stays byte-green. CHANGE-REQUEST v1→v2 at BUILD: v1 gated ALL milestones → the new-milestone scaffold ships the placeholder, so `test_min_pillar`'s census `milestone-confirm mvp` (a no-key scaffold) broke; v2 scoped the gate to opted-in milestones, containing the ripple to test_confirm_parent's 2 opted-in tests (now fill contracts first — in §5 scope). validate-then-write — refuse BEFORE the confirmed:true write. risk:high · conservative gate · §3 FROZEN @ v2 (approved by Tin Dang). red suite test_contract_fill_gate.py [7 tests]. add.py logic delta → ENGINE_MD5 9258dcc7 → 428ca1d1, re-synced across all 3 copies; full suite green. PRIOR re-aim @ confirm-parent (lean-pass M3 flow-simplification 2/3: the OPT-IN confirm-parent gate — `new-milestone --await-confirm` seeds the milestone confirmed:false; WITHOUT the flag NO `confirmed` key is written → grandfathered-confirmed → no gate (so the 342 existing engine tests stay byte-green untouched, ZERO churn); `cmd_new_task` HARD-BLOCKS (validate-then-write, before any scaffold) on `not _milestone_confirmed(state, milestone)` → `_die("milestone_unconfirmed: confirm it first — add.py milestone-confirm <slug>")`; `add.py milestone-confirm <slug> [--by]` is the human gate (validate-then-write — unknown→unknown_milestone; sets confirmed:true/at/by/actor; re-confirm idempotent note; engine RECORDS, never self-confirms). NEW pure `_milestone_confirmed(state, mslug)` mirrors `_setup_locked` one level down (True if confirmed is True OR key absent; False only if confirmed is False). cmd_new_milestone seeds the confirmed/confirmed_at/confirmed_by trio ONLY under `--await-confirm` (plain-create record + output byte-unchanged); `--await-confirm` flag + `milestone-confirm` subparser added. The v1 DEFAULT-ON design (every new milestone gated) was REJECTED at the v1→v2 change-request: its true cost was 342 broken test methods across 61 files (mixed call-shapes, not cleanly scriptable), ~7× the freeze estimate, for no gain over opt-in (the process lesson is about the GUIDED draft-then-confirm flow, where the skill passes the flag). risk:high · conservative gate · §3 FROZEN @ v2 (opt-in, approved by Tin Dang). red suite test_confirm_parent.py [10 tests]. add.py logic+parser delta → ENGINE_MD5 3f82050f → 9258dcc7, re-synced across all 3 copies; full suite green (a verify-gate adversarial refute-read also corrected one stale "(default-on)" inline comment in cmd_new_task → "(OPT-IN)", comment-only, no behavior change). PRIOR re-aim @ identity-in-status (user-identity foundation 3/3 — FINAL, team-collaboration milestone 2: SURFACE the git-native actor — READ-ONLY presentation over task-1's resolver + task-2's stamps, no new writes. (1) cmd_status human gains an `actor   :` line (after `project autonomy:`) = `_whoami(state)` rendered `<name> [<email>] (source: <source>)`, ALWAYS present (resolver TOTAL). (2) `status --json` gains an `"actor"` key = `_whoami(state)` [ratified into test_wave_status_hint's frozen sanctioned set, mirror of active_milestones/active_tasks]. (3) `report_data` task rows gain `gate_actor` [t.get, None when unstamped] + the milestone dict gains `done_actor` — the JSON report facts. (4) `render_report` shows the recorded actor present-only: a ` closed by <name>` header line when done_actor is set + a ` GATED BY` block listing each task's gate_actor; NEW `_fmt_actor(actor)` helper [name <email>, "" when absent]. DESCRIPTIVE + READ-ONLY: no save_state, no new state key, no decision change; an unstamped/legacy record renders NO actor (never a placeholder/crash). FREEZE seam still excluded (carried §7 delta from actor-stamping). risk:high · conservative gate · §3 FROZEN @ v1 (auto-mode standing authorization). red suite test_identity_in_status.py [5 tests] + the sanctioned-set census ratification. add.py logic delta → ENGINE_MD5 d8b9c699 → 04c83b4a, re-synced across all 3 copies; full suite green. PRIOR re-aim @ actor-stamping (user-identity foundation 2/3, team-collaboration milestone 2: STAMP the structured actor on the FOUR engine-WRITTEN human seams, ALONGSIDE today's free-text. NEW `_actor_stamp(state)` = `_whoami(state)` [the SINGLE source; TOTAL — always a non-empty name, so a stamp never fails/blocks a seam] + `_render_actor_line(state)` [`<name> [<email>] (<source>)` for the ledger row]. SEAMS: cmd_lock writes `state["setup"]["actor"]` alongside the existing locked_by free-text · cmd_gate writes `state["tasks"][slug]["gate_actor"]` on EVERY recorded outcome (PASS·RISK-ACCEPTED·HARD-STOP — who recorded the verdict) · cmd_milestone_done writes `state["milestones"][slug]["done_actor"]` · cmd_release adds an `actor:` line to the RELEASES.md row via `_render_releases_row(...,actor=_render_actor_line(state))` [still NO state.json write]. DESCRIPTIVE + ADDITIVE: new optional fields, absent on legacy records → behaves as today; NO existing command's decision reads the stamp; the actor field rides the EXISTING save_state/_atomic_write paths (no new IO, no new failure mode). FREEZE seam EXCLUDED (no engine write — human-authored TASK.md text; seeded as a §7 delta for a future `add.py freeze` command — human-confirmed). risk:high · conservative gate · §3 FROZEN @ v1 (auto-mode standing authorization). red suite test_actor_stamping.py [9 tests]. add.py logic delta → ENGINE_MD5 6f28abab → 610fae59, re-synced across all 3 copies; full suite green. PRIOR re-aim @ actor-identity (user-identity foundation 1/3, team-collaboration milestone 2: the git-native ACTOR resolver — the engine's FIRST git call, strictly fail-soft. NEW `_git_config(key)` [shutil.which gate · subprocess.run(["git","config","--get",key], timeout=2) · catch OSError/SubprocessError/ValueError · empty→None — NEVER raises/hangs, fixed argv, no shell] + `_os_user()` [getpass.getuser() floored, catches KeyError/OSError → "unknown" sentinel] + `_whoami(state)->{name,email,source}` [priority: actor_override (non-blank name)→'override' · git config user.name/email→'git' · _os_user()→'os' floor; TOTAL, name always non-empty, email may be None] + `cmd_whoami` [no-flags show `actor : <name> [<email>] (source: …)` + --json · --name/--email store actor_override (email ""→None) · --unset clear; validate-before-write: reject no_actor_override · actor_name_blank] + the `whoami` subparser (--name/--unset mutually exclusive at the parser → argparse exit 2) + imports shutil/subprocess. NEW optional state key `actor_override`={name,email} (absent by default → behaves as today). DESCRIPTIVE only, no existing command's decision changes. risk:high · conservative gate · §3 FROZEN @ v1 (auto-mode standing authorization). LAYERED ON the fold-suggestion-seams base (1911b250) after a live parallel-writer contention reconciliation (the fold session committed feat/fold-suggestion-seams @ 9dc3e944 = PR #47, clobbering the in-flight actor-identity engine in the shared tree; re-applied additively here). A verify-gate independent review [python-expert, BLOCK, 0.97] had found TWO real crash paths violating the frozen fail-soft/TOTAL invariants, both fixed before this re-application: (1) `_git_config` now catches UnicodeDecodeError [a ValueError subclass — a non-UTF-8/latin-1 git config value would otherwise crash the engine's first git call]; (2) `_os_user` floors getpass.getuser()'s KeyError [bare container: no passwd entry + no LOGNAME/USER] → "unknown" so _whoami stays TOTAL. Plus 2 NITs: --email "" normalizes to None; --name/--unset mutually exclusive. red suite test_actor_identity.py [17 tests: + test_failsoft_nonutf8_value · test_os_floor_total_when_getuser_raises · test_set_empty_email_normalized_to_none · test_name_and_unset_mutually_exclusive; test_failsoft_nonzero_or_empty → test_failsoft_empty_stdout] + test_min_pillar census (LIFECYCLE += `whoami`). add.py logic+parser delta → ENGINE_MD5 1911b250 → 6f28abab, re-synced across all 3 copies; full suite green. PRIOR re-aim @ parallel-status-view (team-collaboration foundation 4/5: cmd_status renders the active SET as PARALLEL STREAMS — presentation-only, no command DECISION changes. THREE additive edits: (1) the human-readable branch prints a `streams : <N> active milestones` block AFTER the unchanged `active  :` line, gated on len(active_milestones)>=2, one line per active milestone [PRIMARY first via [_primary]+[m!=_primary], primary marked `▸` + tagged `(primary)`] naming task=active_tasks[m] or (none) and phase=tasks[tk].phase or `-`; (2) the milestone-rollup mark becomes `"*" if mslug in active_milestones else " "` [N<=1 identical since active_milestones==[primary] or []; N>=2 marks every active member]; (3) `status --json` gains additive keys active_milestones [list] + active_tasks [map], every existing key UNCHANGED. N<=1 text + existing json byte-for-decision identical — the full suite is the untouched oracle [the standing additive-cue convention shared with the graduation/release/spec/fold cues]. risk:high · conservative gate · §3 FROZEN @ v1 (auto-mode standing authorization). add.py logic delta → ENGINE_MD5 929ced7e → fa8e9818, re-synced across all 3 copies; red suite test_parallel_status_view.py [8 tests], full suite green. PRIOR re-aim @ multi-active-commands (team-collaboration foundation 3/5: the multi-active COMMANDS that USE the task-2 seam — NEW _activate_milestone(state, slug) [append-to-SET-if-absent + focus the scalar mirror + repoint scalar active_task from active_tasks[slug]] · _deactivate_milestone(state, slug) [remove-from-SET + pop active_tasks[slug]; if it was the primary, repoint to the SET's last member, else leave the primary untouched]. NEW verbs: `activate <slug>` [cmd_activate: reject unknown_milestone if slug∉milestones, milestone_done if status==done, then _activate_milestone] · `deactivate <slug>` [cmd_deactivate: reject milestone_not_active if slug∉active_milestones, then _deactivate_milestone] — both validate-before-mutate (a reject leaves state byte-unchanged). cmd_use made MILESTONE-AWARE: focuses the task's OWN milestone (_activate_milestone(ms) + _set_active_task(state, slug, ms)) when the task carries a known milestone, else scalar-only (back-compat) — closing the task-2 nit-2 §7 delta (cross-milestone map entry). cmd_archive_milestone's clear-pair replaced by _deactivate_milestone(state, slug) (drop-from-SET + pop the task entry, repointing primary). new-milestone keeps its REPLACE-to-focus (additive activate/deactivate are the SET writers; task-2's frozen _set_active_milestone untouched). risk:high · conservative gate · §3 FROZEN @ v1 (auto-mode standing authorization). add.py logic+parser delta → ENGINE_MD5 5eb79fa8 → 458fb992 → 929ced7e, re-synced across all 3 copies; red suite test_multi_active_commands.py [13 tests], full suite green. v2 (929ced7e) = a verify-gate independent review [python-expert, BLOCK, 0.97] found a NON-primary-archive N≤1 regression: routing cmd_archive_milestone solely through _deactivate_milestone DROPPED the task-2 guard that cleared the scalar active_task when the archived milestone's MEMBER task was the scalar focus while the PRIMARY was a different milestone (reachable via new-milestone replace-to-focus, which leaves the scalar stale) — leaving active_task dangling at a now-deleted task (phantom unknown_task for slug-less advance/gate). Fixed by restoring the N≤1 guard AFTER the SET-level deactivate: `if _active_task(state) in members: state["active_task"]=None` (the map entry is already popped by _deactivate_milestone; the frozen archive-routes-through-_deactivate clause stands, the back-compat scalar clear is additive) + a red→green guard test_archive_nonprimary_clears_stale_scalar_task. Review NOTE (not blocking, contract-faithful): cmd_use re-activates a DONE milestone (no done-check, unlike cmd_activate) → seeded as a §7 delta. PRIOR re-aim @ active-accessors (team-collaboration foundation 2/5: the active milestone/task ACCESSOR SEAM — NEW _active_milestone(state) · _active_task(state, milestone=None) · _set_active_milestone(state, slug) · _set_active_task(state, slug, milestone=None) [reads return the N≤1 scalar mirror; writes keep scalar + active_milestones/active_tasks in sync; orphan active-task = scalar-only; clear pops the map entry] + EVERY active-milestone/active-task read & write site routed through them (incl. cmd_check's formerly-direct read — consumes the state-schema-migration §7 deltas: the 3rd seam + active_tasks partial-state default). Behavior-preserving for N≤1: the full suite stays green with NO test changed (the byte-for-decision-identity oracle). risk:high · conservative gate. §3 FROZEN @ v1 (auto-mode standing authorization; [contract] "primary"=scalar today, per-milestone arg exists). add.py logic delta → ENGINE_MD5 aef59d9c → 995d9369 → 5eb79fa8 (v2: a verify-gate independent review [python-expert, MERGE-WITH-NITS, 0 blocking, 0.97] found a stale active_tasks entry left after archive — fixed by clearing the active task with its milestone still known [_set_active_task(state, None, slug) before _set_active_milestone(state, None)] so active_tasks[slug] is popped; nit 2 [cmd_use cross-milestone map entry, pre-existing] → §7 delta for multi-active-commands), re-synced across all 3 copies; red suite test_active_accessors.py [10 tests], full suite green) · re-aimed @ state-schema-migration (team-collaboration foundation 1/5: the multi-active state schema + a PURE/idempotent/TOTAL `_migrate_state(state)` forward-migration wired into BOTH load seams — load_state + _load_state_for_json — that grandfathers a single-active state to the new shape on load: adds `active_milestones` [the active SET, derived from the scalar active_milestone: None→[], "x"→["x"]] + `active_tasks` [per-milestone active-task MAP]; KEEPS the scalar active_milestone/active_task as the N≤1 mirror so the ~20 not-yet-routed readers keep working; an orphan active_task [one belonging to no active milestone] stays a top-level scalar fallback [FROZEN decision (a)]; an already-migrated state is returned unchanged [idempotent]; corrupt JSON still dies state_invalid in the loader [migration never reached]. init is BORN-MIGRATED [active_milestones:[] + active_tasks:{} written natively]. risk:high · conservative gate. §3 FROZEN @ v1 ([contract] orphan=top-level fallback · storage=active_tasks map · scalar mirrors kept). add.py logic delta → ENGINE_MD5 25c6fcc5 → 8a914147 → aef59d9c (v2: a verify-gate independent review [python-expert, MERGE-WITH-NITS, 0 blocking] flagged the "PURE" docstring/contract word against the in-place mutation — fixed so the helper upgrades on a fresh top-level `dict(state)` copy and NEVER mutates the caller's dict; behavior otherwise identical, 11 tests stay green), re-synced across all 3 copies; red suite test_multi_active_state.py [11 tests], full suite green) · re-aimed @ scope-walk-artifacts (exclude gitignored build artifacts — `.next`/`coverage`/`test-results` dirs + `.pyc`/`.tsbuildinfo` suffixes — from the §5 scope-walk, alongside the existing `.serena` cache; supersedes the fold-command pin) · re-aimed @ fold-command (delta-resolution 4/4: NEW `add.py fold` — the HUMAN-AUTHORIZED reversal of the frozen "engine stays judgment-free; there is no add.py fold" principle (test_foundation_update_loop re-frozen @ v3). Mechanizes ONE competency-lesson consolidation session, TRANSCRIPTION-ONLY: it routes each lesson's OWN captured text into its routed home and NEVER composes/merges prose (consolidation judgment stays the human's, via the compaction door). cmd_fold [--task/--comp] + the `fold` subparser + PURE _fold_competency_delta(text, version, comps) [flips EVERY open `[<COMP> · open]` → `folded` + ` [folded foundation-version N]` stamp via _COMP_OPEN_TOKEN_RE, mirror of _resolve_spec_delta; line text + `(evidence: …)` byte-preserved; None when nothing open] + _FOLD_ROUTES {DDD/SDD/UDD → PROJECT.md §Domain/§Spec/§Users · TDD/ADD → CONVENTIONS.md §Method learnings, TOTAL over the five} + _section_present/_prepend_to_section/_prepend_key_decision_row. BATCH = one call → ONE foundation-version bump → every stamp carries the SAME N (fold.md "one session, one bump"). VALIDATE-ALL-THEN-WRITE: ≥1 open lesson in scope · every routed section present (incl. §Key Decisions) · parseable `foundation-version:` — all checked + every new body built in memory BEFORE any _atomic_write, so a reject leaves the WHOLE tree byte-unchanged. 3 reject codes: no_open_deltas · missing_route_section · no_foundation_version. `fold`/`folded` are Group C machine tokens referenced by NAME (_FOLD_VERB/_FOLDED) inside output strings so the ubiquitous-language prose lint sees no slang. AUTHORIZED guard change-requests: test_foundation_update_loop @v2 → @v3 (invert no-fold-command → the command EXISTS + transcription-only; REJECT_CODES → the new 3) · test_min_pillar census (LIFECYCLE += `fold` · _NONZERO_OK += `fold` — refuses no_open_deltas on a delta-free board) · test_ubiquitous_language MACHINE_CONSTANTS += `fold`. fold.md rewritten ×3 to describe the command (prose stays "consolidation"; `fold`/`folded` in code spans only). risk:high · conservative gate. §3 FROZEN @ v1 ([contract] transcribe-not-consolidate + uniform `(<COMP>)`-prefixed verbatim bullet, human-approved at the freeze). VERIFY-gate refute-read hardened the cross-file commit: cmd_fold now stages EVERY edit to a temp then renames all via the NEW _atomic_write_many(writes) two-phase helper, so a mid-commit IO failure (disk-full/permission) leaves NOTHING written — the frozen §3 "any failure → write nothing" now holds across the N task-files + PROJECT.md + CONVENTIONS.md set, not just per-file; writes ordered foundation-FIRST so the near-impossible mid-rename residual is a still-open lesson (visible, re-runnable) never a flipped-but-untranscribed one (silent loss). Success + precondition-reject paths byte-identical (7 fold tests stay green). A SECOND VERIFY-gate finding (dogfooding HARD-STOP, human-authorized to fold into THIS task): the §5 scope guard's `_scope_walk` excluded only `.git/.add/__pycache__/node_modules` — NOT `.serena`, serena's code-intelligence symbol-index cache, which re-writes itself whenever a source file changes. So the AI's OWN add.py edits churned `.serena/cache/*.pkl`, the build-entry snapshot baked them in, and the gate flagged them as out-of-scope touches (un-fixable at gate time: deletion reads as a delete-touch, regeneration as a modify-touch) — exhausting the heal loop to a FALSE-POSITIVE HARD-STOP. Fix: `.serena` added to `_SCOPE_EXCLUDE_DIRS` (a tool cache with no build signal, the exclude set's exact purpose; widening it is the sanctioned additive change-request the constant's own comment names). Proven red/green: pre-fix `.serena/...pkl` leaks into the snapshot, post-fix pruned — guarded by a new assertion + fixture in test_scope_gate_enforce.py (added to §5 scope). The task was reopened → re-crossed tests→build so the FRESH snapshot excludes `.serena` from birth, making the gate immune to serena re-indexing. A pre-MERGE independent review (python-expert, verdict MERGE-WITH-NITS, zero blocking) flagged ONE faithfulness nit: the cmd_fold inline commit comment overstated atomicity ("leaves NOTHING written" is true only for the phase-1 temp-write failure; a phase-2 mid-rename residual re-runs by DUPLICATING, not cleanly). Comment-only honesty fix — no behavior change (the two-phase commit + foundation-first ordering stand). add.py logic+parser delta → ENGINE_MD5 392b4e55 → 58ad75cb → b65fc266 → f78a2f24 → 5f3f7d28, re-synced across all 3 copies; red suite test_fold_command.py [7 tests] + test_scope_gate_enforce .serena guard, full suite green). Supersedes re-aimed @ spec-delta-guards (delta-resolution 3/4: SPEC-delta GUARDS make an open SPEC delta impossible to lose silently — PROJECT-WIDE (human change-request from member-scoped at the freeze; report unified too). cmd_compact gains a `open_spec_deltas_unresolved` HARD guard AFTER open_deltas_unfolded, BEFORE the first rename (validate-before-move; a reject leaves tree+state byte-unchanged) — refuses while ANY task in the project holds an open SPEC delta, deliberately broader than the member-scoped competency guard (a SPEC delta is a project-global forward hand-off). Three READ-ONLY nudges, all on the SAME source `len(_collect_open_spec_deltas(root))`: cmd_status `spec : N open SPEC delta(s) — resolve: new-task --from-delta / drop-delta` (sibling of the `deltas : N open` fold nudge) · cmd_milestone_done `note: N open SPEC delta(s) to resolve (seed/drop)` (sibling of the competency fold note) · report_data.summary[open_spec] + a render ` SPEC DELTAS  N open` line. NO new collector (reuses the task-1 reader), NO new subcommand (census untouched — task-2 verb-vs-flag lesson). add.py logic delta → ENGINE_MD5 5120b0c2 → 392b4e55, re-synced across all 3 copies; red suite test_spec_delta_guards.py [5 tests], full suite green. Supersedes re-aimed @ seed-and-drop (delta-resolution 2/4: NEW SPEC-delta resolution VERBS — the first WRITERS of the seeded/dropped statuses task 1 only TOLERATED on read. (1) `new-task <new> --from-delta <prior>` SEEDS the prior's first open SPEC delta into the new task — reads it via _first_open_spec_text, pre-fills §1 `Feature: <text> (from <prior> spec-delta)`, flips the source `[SPEC · open]` → `[SPEC · seeded] … [→ <new>]`, records `state[tasks][new].from_delta=<prior>` lineage. (2) `drop-delta <task>` DISMISSES the first open SPEC delta → `[SPEC · dropped]`. NEW pure _resolve_spec_delta(text, new_status, pointer=None)->str|None [flips ONLY the first open line's status token via _SPEC_OPEN_TOKEN_RE + optional ` [→ ptr]` stamp; entry text + `(evidence: …)` byte-preserved; None when no open delta → caller refuses, no write] + _first_open_spec_text [first open delta's text, evidence stripped, from the SAME in-memory text the flip uses] + cmd_drop_delta + the `drop-delta` subparser + `--from-delta` on new-task/cmd_new_task. DESIGN-FOR-FAILURE: validate-ALL-then-write — resolve prior · prove ≥1 open SPEC delta · compute the seeded flip BEFORE any write; per-file _atomic_write; ordered new-TASK.md → prior-flip → state so the worst residue is 'new task created, prior still open & visible' [re-run blocked by the slug-exists guard → never a dangling [→] pointer at an absent task]. Additive verbs — no existing test enumerates the subcommand set, so no downstream test touched. §3 FROZEN @ v1 ([spec] first-open selector + [contract] no cross-file txn, both human-approved at the freeze). add.py logic+parser delta → ENGINE_MD5 7d908b2a → 5120b0c2, re-synced across all 3 copies; red suite test_seed_and_drop.py [8 tests], full suite green). Supersedes re-aimed @ spec-delta-grammar (delta-resolution 1/4: NEW SPEC-delta resolution track — a SEPARATE delta lifecycle from the five competency deltas. New grammar constants beside _DELTA_RE (single-source, no parallel grammar): _SPEC_STATUSES=(open,seeded,dropped) · _SPEC_DELTA_RE · _STATUS_SETS{tag→legal-status-set}. _spec_delta_entries + _collect_open_spec_deltas(root)->[{task,text,evidence}] (FLAT, open-only; SPEC resolves into a TASK, never consolidated into the foundation, so collected APART from _collect_open_deltas). _task_prose observe now derives first-open-SPEC → legacy 'Spec delta for the next loop:' field (back-compat) → '(unknown)'. _lint_task_deltas now lints BOTH '### Competency deltas' AND '### Spec delta', TAG-SCOPED via _STATUS_SETS: a cross-set status ([SPEC·folded], [SDD·seeded]) → unknown_status; evidence required on an OPEN entry of ANY tag; resolved-for-tag (folded|rejected · seeded|dropped) skipped (unknown tag defaults to the competency set so legacy folded lines still skip). cmd_deltas gained an 'open spec deltas' section + a separate --json 'spec'/'spec_total' key. TASK.md.tmpl §7 + _FALLBACK_TASK: the free-text 'Spec delta for the next loop:' line REPLACED by a '### Spec delta' block (no net new template comment — lean-pass <12 held). §3 FROZEN @ v1 (SPEC evidence REQUIRED, symmetric with competency — human-chosen at the freeze). add.py logic+parser delta → ENGINE_MD5 9c479010 → 7d908b2a, re-synced across all 3 copies; red suite test_spec_delta_grammar.py [11 tests]; downstream test_report observe tests + the template re-synced; full suite green. Supersedes re-aimed @ release-command (release-altitude 3/4: NEW GUARDED, record-only `add.py release <version>` — the ACT half of the 5th scope level (RELEASE, release.md), sibling to the release-report GATHER. cmd_release enforces the 4-code readiness floor THEN records the cut: release_security_open (≥1 open HARD-STOP gate — the UN-FORCEABLE reject, `--force` is ignored: a security finding is never shipped) · release_tests_red (an in-flight build w/o a recorded green gate — the tool-agnostic PROXY _build_in_flight(state): any active task phase∈{build,verify} AND gate==none, since the engine never shells to a test runner; `--force` overrides) · release_no_closed_milestone (_releasable==[]; `--force` overrides) · release_undisclosed_waiver (release_data['waivers'] non-empty AND no --with-waivers; `--force` overrides). VALIDATE-BEFORE-WRITE: all floor checks run before the first write so a reject leaves both files + state.json byte-for-byte unchanged. RECORD = exactly TWO writes via _atomic_write at the PROJECT ROOT (root.parent): CHANGELOG.md (prepend a `## <ver> — <date>` block, create w/ `# Changelog` header if absent) + RELEASES.md (prepend a newest-first row `## <ver> — <date>`·`milestones:`·`waivers:`·`evidence:`). The row's `milestones:` line IS the attribution (task-2 decision) — so cmd_release writes NO state.json (the `→ releasable` cue re-reads the ledger). Engine RECORDS, human SHIPS: NO git tag/publish/deploy, NO version-source bump. DESIGN-FOR-FAILURE: contents built in memory first; CHANGELOG written, then RELEASES.md (the commit point); a failed 2nd write ROLLS BACK CHANGELOG to its prior bytes (or unlinks it if created) and refuses release_write_failed — never a half-recorded cut. New helpers _build_in_flight · _prepend_block (newest-first, append-only — existing rows preserved verbatim) · _render_changelog_block · _render_releases_row; new `release` subparser (version positional + --force + --with-waivers + --evidence). test_min_pillar census: LIFECYCLE += `release 0.0.0` (refuses release_no_closed_milestone at its slot) + `release` ∈ _NONZERO_OK. §3 FROZEN @ v1 (risk:high, conservative gate). add.py logic+parser delta → ENGINE_MD5 d0d9b1ed → 9c479010, re-synced across all 3 copies; red suite test_release_command.py [12 tests], full suite green. Supersedes re-aimed @ release-report (release-altitude 2/4: NEW read-only `add.py release-report` — the GATHER half of the 5th scope level (RELEASE, release.md). release_data(root, state) clusters FIVE record-sets {releasable · changed · waivers · blockers · monitors + summary}: PURE·NO writes·exit 0 ALWAYS (only no_project non-zero)·text default + --json frozen facts interface — mirrors graduation_data/cmd_graduation_report. GATHER-not-JUDGE by construction: no readiness/score/ranking/verdict field (the FORBIDDEN-token guard). NEW helpers: _releases_path (the append-only ledger at the PROJECT ROOT = root.parent, a sibling of CHANGELOG.md — NOT inside .add/) · _released_milestones (the union of every `milestones:` row; FAIL-OPEN — a missing/unreadable/malformed ledger → empty set → every closed milestone reads still-releasable, so a vanished ledger never hides work) · _closed_milestones (LIVE status==done + ARCHIVED, all PASS-done) · _releasable (closed MINUS released — drives BOTH the status cue and the report) · _key_decisions_for (best-effort §Key-Decisions rows from PROJECT.md naming the slug, fail-open []). NEW module constants RELEASABLE_CUE + RELEASES_FILE. cmd_status gains the project-global read-only cue `→ releasable: N milestone(s) closed since last release` (fires for a live-done OR archived unreleased milestone — stands with no live milestones; additive; ledger read fail-open so a vanished ledger never silences it). NEW `release-report` subparser (--json/--plain) registered after graduation-report. Attribution source = RELEASES.md MEMBERSHIP (a milestone is released iff its slug is on a `milestones:` row) — chosen so the cue never reads compacted milestone files (the flagged riskiest decision). §3 FROZEN @ v1 (conservative gate). add.py logic+parser delta → ENGINE_MD5 91e76c35 → d0d9b1ed, re-synced across all 3 copies; red suite test_release_report.py [8 tests], full suite green. Supersedes re-aimed @ capture-evidence (udd-design-loop 3/4: NEW pure _missing_captures(root) -> list[str] [prototype names under .add/design/prototypes/ with no .add/design/captures/<name>.<ext> for ext in png·svg·jpg·jpeg·webp; PURE·TOTAL·READ-ONLY·doc-order; [] when proto_dir absent] + cmd_check appends a never-red WARN missing_capture per uncaptured prototype on the EXISTING `warnings` array — NEVER `checks`, so never feeds `failed`; silent-when-absent; the engine MEASURES capture presence, never renders. design.md beat 4 + "## Tool-agnostic capture" reopened ×3 skill trees + udd-wireframe.md "## Capture is evidence" ×2 template trees: captures at .add/design/captures/<name>.<ext> · attached/mentioned in the feature's TASK.md · @json-render/image [Satori→PNG/SVG] the named-default capture engine. Worked welcome/settings captures committed to .add/design/captures/ + cited in §6 as the demonstration [exit-criterion 4]. add.py logic delta → ENGINE_MD5 5599740d → 91e76c35, re-synced across all 3 copies; red suite test_capture_evidence.py [14 tests], full suite green. Supersedes re-aimed @ agent-update-nudge (1.4.0 release — integrate fork PR #17 [ewalliss]: the agent-agnostic stale-install nudge. ADD tells EVERY agent to run `add.py status`/`guide` first each session, so that is the one universal chokepoint to flag a stale install. main() gains a fail-open _maybe_nudge_update(args) preamble — STDERR-only, NEVER touches stdout or the exit code, gated to cmd_status/cmd_guide ONLY: it reads the launcher .add-version stamp [version+channel] and, when the registry latest > current, writes an ACTION-REQUIRED line naming the channel-correct command [npx @pilotspace/add@latest update · pipx run pilotspace-add update]. The engine's ONE deliberate network touch, bounded: 1.5s timeout, throttled once/24h via .update-cache.json, fail-open on offline/timeout/bad-payload [stale cache beats nothing], INERT without a launcher stamp [engine init writes no stamp, so the offline suite never touches the network], kill switch ADD_NO_UPDATE_CHECK=1. New helpers _read_json_safe · _write_json_safe · _version_gt · _fetch_latest_version · _cached_latest; imports += urllib.request, datetime += timedelta. IMPROVED over the raw PR: .update-cache.json added as a 3rd _GITIGNORE_BODY entry [transient-artifact hygiene] + new tests covering the cache-write & stale-cache-fallback branches. add.py delta → ENGINE_MD5 6c5ba081 → 5599740d, re-synced across all 3 copies; red suite test_update_nudge.py + extended test_add gitignore guards, full suite green. Supersedes re-aimed @ soul-artifact (v13-onboarding-polish 5/6: SOUL.md — a NEW survivor-layer living "voice" doc named Trusting — joins the foundation. add.py delta is TWO additive seams: SETUP_FILES += "SOUL.md" (last member) so cmd_init scaffolds .add/SOUL.md via the GENERIC render+write loop [never-clobber an existing one · skip-if-blank warns rather than write a 0-content survivor]; + a cmd_status VOICE pointer ["voice   : .add/SOUL.md (how I sound & what keeps your trust — read each session)"], existence-gated [no open/parse → no new IO failure path], sibling to the PROJECT.md context pointer. Ships templates/SOUL.md.tmpl [×3 byte-identical via test_bundle_parity + the .add dogfood copy] carrying the SOUL SCHEMA — Name · Tone · Communication style · Trust · Learns-from · Voice deltas — with a PROPOSED "Trusting" starter voice EXPLICITLY marked human-owned + overridable [the identity-values-human-owned guideline: tests assert the SCHEMA/mechanism ONLY, never the tone words, so the human rewrites the voice with zero build breakage] + a pointer to the soul-self-improve loop as the only voice-delta writer. + the skill SKILL.md "Always start here" gains the read-.add/SOUL.md-each-session instruction [skill ×3, test_tree_parity + test_bundle_parity]. test_add.test_init_creates_state_and_setup_files auto-covers SOUL.md [it iterates SETUP_FILES]. §3 FROZEN @ v1 [autonomous authorization; voice CONTENT surfaced OPEN to the human at the seam — soul-self-improve task 6 owns convergence to the human's wording]. add.py SETUP_FILES+cmd_status delta → ENGINE_MD5 eebbb443 → 6c5ba081, re-synced across all 3 copies; new red suite test_soul_artifact.py [8 tests], full suite green. Supersedes re-aimed @ dag-scheduler (v13-onboarding-polish 1/6: NEW read-only `add.py waves` DAG scheduler — _wave_schedule(state, mslug) [PURE·TOTAL: never mutates, never raises on dict input] + cmd_waves + the `waves` subparser. REUSES cmd_ready's satisfaction predicate [a dep is satisfied iff archived OR _task_done] and the _find_cycle idiom; groups the active milestone's not-done tasks into topological Kahn waves [wave = tasks whose open-member deps all landed earlier], names the CRITICAL PATH [longest member-dep chain, ties→smallest slug] via memoized depth, emits an advisory TIER HINT [top on the critical path, mid elsewhere — a proxy for scope-of-impact; run.md model pick stays advisory, never a gate], and surfaces a BLOCKED set [a not-done NON-member dep is unsatisfiable within this milestone → reported with waiting_on, never silently dropped]; an all-done/empty milestone schedules nothing [exit 0], the rejects are no_active_milestone · unknown_milestone · dependency_cycle [names the cycle members]. READ-ONLY: no save_state, no `next:` footer [read-only verb]. --json emits {milestone, waves, critical_path, critical_path_len, tiers, blocked}. streams.md gains the "## DAG strategy — let the engine schedule the waves" section [×3 byte-identical]. §3 FROZEN @ v1 [autonomous authorization]. add.py logic+parser delta → ENGINE_MD5 475d9d40 → 02426e9d → eebbb443 [v1.1: the VERIFY adversarial refute-read found a real HIGH bug — blocking did not propagate transitively, so a task whose only dep was a BLOCKED sibling was mis-placed in wave 1 as ready; fixed by computing the blocked partition to a FIXED POINT, +1 red guard test test_transitive_blocked_dep + 4 gap tests (empty milestone · self-dep cycle · linear-chain critical path · read-only across invocations)], re-synced across all 3 copies; red suite test_dag_scheduler.py [14 tests], full suite 1058 green. Supersedes re-aimed @ autonomy-command (NEW first-class `add.py autonomy show|set` verb — the autonomy level finally gains the CLI verb it lacked, so an agent under `auto`, applying the correct "first-class state has a command" model, can no longer hallucinate a phantom `add.py autonomy` and derail [the scla-mono symptom that opened this task]. cmd_autonomy + the `autonomy` subparser [positional action show|set · a1/a2 · --project · --yes; _opt_positionals=(a1,a2)]: `show` is read-only — prints declared · effective[fallback-resolved] · project default · the verify-gate owner [the SAME _driver_stop the footer/guide render]; `set` is the FIRST writer of the `autonomy:` header token — _autonomy_decl_line replaces the SINGLE declaration line idempotently [trailing comment preserved, NEVER appended — kills the scla-mono duplicate-line bug; inserts after slug:/heading when absent], _guard_autonomy_raise refuses a RAISE toward auto without --yes [autonomy_raise_unconfirmed — the engine's first confirm flag, the human-owned trust escalation], the risk:high guard [unguarded_high_risk_auto, reused _RISK_HIGH_RE + lowered-rung] runs BEFORE the write, a bad level → autonomy_level_invalid; reused VERBATIM _require_root + _resolve_task; state.json UNCHANGED [autonomy stays a header token]. De-command-shape: the 3 command-shaped "set autonomy: X" strings [garbled-project warning · the gate risk-guard · the check implicit_autonomy warn] + templates/TASK.md.tmpl now cite `add.py autonomy set`; a WORDING_RUBRIC.md idiom_map [enforced] fence bans the command-shaped idiom so it can never regress on the agent surface. §3 FROZEN @ v2 [v1→v2 change-request raised at TESTS: reject codes aligned to the guards it reuses]. add.py logic+parser delta → ENGINE_MD5 f15eb52a → 475d9d40, re-synced across all 3 copies; new red suite test_autonomy_command.py [16 tests] + sibling census ratifications [test_min_pillar LIFECYCLE += autonomy · test_rewrite_guides AUTONOMY_COMMAND_IDIOMS], existing engine suites stay green). Supersedes re-aimed @ gitignore-scaffold (vcs-hygiene: cmd_init now scaffolds a co-located .add/.gitignore — module constant _GITIGNORE_BODY ignoring scope-snapshot.json + pre-archive-state.bak.json (bare-filename patterns, any depth under .add/); written after the tasks/ mkdir, NEVER-CLOBBER an existing copy, additive — touches no snapshot/.bak on disk. add.py-only delta → 7b67afd1 → f15eb52a, re-synced across all 3 copies; guards: test_add.test_init_scaffolds_gitignore + test_gitignore_ignores_transient_artifacts (real git check-ignore enforcement) + test_init_never_clobbers_existing_gitignore. Supersedes re-aimed @ udd-check-lint (udd-design-foundation task 4/4: the COMPOSER — wire BOTH pure validators into `add.py check` + close the cross-file token resolution tasks 1+2 deferred. NEW cmd_check "UDD foundation" section (read-only, appended before the tally): discovers the named set under `.add/design/` (Fork A, human-chosen at the freeze over the `.add/` root) — root/design/tokens.json · root/design/catalog.json · root/design/prototypes/*.json; FAIL-CLOSED load (malformed JSON → malformed_tokens_json · malformed_catalog_json · malformed_prototype_json, never _die, never a crash); composes _token_layer_violations(tokens) + _catalog_tree_violations(catalog, tree) UNCHANGED, one FAIL per violation (named code as reason) / one PASS per clean file; SILENT-WHEN-ABSENT (no named set → zero new output, so clean/non-UI projects + this dogfood repo stay green). NEW pure _prop_token_resolution_violations(tokens, catalog, tree) (+ helper _semantic_token_index) closes the deferral: resolves every tree token-prop alias targeting `semantic` against tokens.json — terminal node absent / not a token → unresolved_prop_token; the resolved semantic token's INHERITED $type (DTCG: $type on the GROUP, leaf carries only $value) ≠ the catalog PropSpec's `token:<$type>` → prop_token_type_mismatch. PURE + TOTAL (never mutates, never raises on dict inputs); acts ONLY on a catalog token-PropSpec × a `{semantic.*}` tree alias (the props task 2 passed LAYER-only) — non-token/non-alias/non-semantic/uncataloged are SKIPPED (task 1/2 own those codes; no double-flag). cmd_check's read-only-ness + output format (PASS/FAIL/WARN · "check: N passed, M failed (K warnings)" · the --json schema) + SystemExit(1)-iff-failed are PRESERVED — the ~30 existing cmd_check tests stay green. Fork A reconciliation (human's freeze choice): the DESIGN.md.tmpl + udd-catalog.md live-file pointers move to `design/…` (×3 each; the task-3 substring tests survive — `design/tokens.json` still contains `tokens.json`). new red suite test_udd_check_lint.py (18): the resolver unit tests (clean · the shipped samples cross-resolve clean · unresolved · $type-mismatch · group-alias · skip-non-token/non-semantic · skip-malformed-PropSpec · skip-malformed-resolved-$type · purity) + the cmd_check integration over .add/design/ (clean-passes-read-only · each named red · fail-closed malformed tokens/catalog/prototype · silent-when-absent). add.py-only logic change → ENGINE_MD5 e2e1dd0 → 7e0ad6a → 7b67afd1 (v2 — the verify-gate adversarial refute found the cross-file resolver DOUBLE-FLAGGED two shapes the §3 'no double-flag — task 1/2 already own those codes' clause forbids: a malformed token PropSpec [bad / absent `token` → task-2's malformed_catalog] and a resolved semantic token whose inherited $type is malformed / absent [→ task-1's unknown_type]. Fix: the resolver now SKIPS a prop whose declared `token` ∉ _TOKEN_TYPES AND skips the mismatch arm when the resolved token's $type ∉ _TOKEN_TYPES — both routed to their task-1/2 owners; the unresolved-alias + legit-mismatch paths are UNCHANGED. +2 red no-double-flag resolver tests + 2 fail-closed integration tests for malformed_catalog_json · malformed_prototype_json [§1 Reject codes that had shipped untested] + the clean test strengthened to assert the prototype PASS line; 14→18 tests). This pin supersedes re-aimed @ udd-design-template (udd-design-foundation task 3/4: the prose FRONT-DOOR — wire a DESIGN.md template into setup. ENGINE delta is ONLY SETUP_FILES += "DESIGN.md" (appended last) so cmd_init renders the NEW templates/DESIGN.md.tmpl via the GENERIC L377–389 render+write loop → .add/DESIGN.md (skip-not-clobber; a blank render is warn-skipped). The template is a STRUCTURED BINDING doc (Fork A) — SIX required sections: a header + a "no UI? this doc is optional — delete it" self-note · ## Identity (brand·palette·typeface·voice as HTML-comment PROMPTS, NO pre-filled brand value — the identity-values-human-owned guideline; a concrete hex outside a prompt = identity_prefilled) · ## Principles/persona · ## Screens (a prompted prototypes/<name>.json index SEEDED with prototype.sample.json — the human-added 6th section at the v1 freeze, "Freeze + Screens section") · ## Foundation (pointers binding the named set tokens.json/catalog.json/prototypes/ + their docs udd-tokens.md/udd-catalog.md + the shipped samples tokens.sample.json/catalog.sample.json/prototype.sample.json, Fork B — start from the samples) · ## Render (a pointer to udd-catalog.md's "## Render recipe"). Rendered with {{project}}/{{stage}}/{{date}}; MUST render non-blank. FIVE named reds (the DESIGN.md guards; udd-check-lint does NOT lint this prose): design_not_wired · design_template_blank · identity_prefilled · design_section_missing · mirror_drift. Fork C: wiring is UNCONDITIONAL — every init drafts DESIGN.md, and the self-note tells a non-UI project to delete it; the engine delta is ONLY the one SETUP_FILES tuple member (the render+write loop is untouched). Ships templates/DESIGN.md.tmpl (canonical+bundle byte-identical via test_bundle_parity, + the .add dogfood copy for the dogfood engine) + the phases/0-setup.md "Draft to the lock" DESIGN.md line (skill ×3, test_tree_parity + test_bundle_parity) + test_init_creates_state_and_setup_files auto-covers DESIGN.md (it iterates SETUP_FILES). add.py-only one-tuple-member delta → ENGINE_MD5 3cdfaab → e2e1dd0. No contract change-request — the §3 freeze added Screens at v1 before the stamp. This pin supersedes re-aimed @ udd-catalog-content-schema (udd-design-foundation task 2/4: the render-ready half — a NEW pure validator _catalog_tree_violations(catalog, tree) -> list[(code, path, detail)] (+ helpers _propspec_malformed, _prop_value_code) checking OUR compact-JSON component CATALOG against a flat json-render content TREE. CATALOG = {components: {<Name>: {description?, hasChildren?(default false), props: {<prop>: PropSpec}}}}; a PropSpec is exactly one of {type: string|number|boolean} · {type: enum, values:[str]} · {type: token, token:<$type>} ($type ∈ the task-1 set color·dimension·number·fontFamily·fontWeight·duration). CONTENT TREE = json-render flat Spec pinned @ vercel-labs/json-render v0.19.0 / commit 4e4dc46 (the milestone's named top risk: young-project schema drift): {root, elements: {<id>: {type ∈ catalog · props ⊆ declared · children? only if hasChildren, ids ∈ elements}}}; a token-prop value is the {semantic.dotted.path} alias; json-render's state/on/visible/repeat are PASSED THROUGH (render-compatible + clickable) but NOT linted, keeping the validator lean. EIGHT named reds: tree_cites_uncataloged_component · unknown_prop · prop_type_mismatch (incl. a token prop given a non-alias literal) · non_semantic_prop_token · dangling_child · children_not_allowed · missing_root · malformed_catalog. PURE (never mutates inputs), stdlib-only, DETERMINISTIC document order; [] == valid. CHANGE-REQUEST raised at TESTS → frozen §3 @ v2 (Option B, human-approved): the frozen (catalog, tree) signature cannot resolve a token's existence/$type (that needs tokens.json), so non_semantic_prop_token was NARROWED to LAYER-only — a token-prop alias MUST target the `semantic` layer (first path segment); target existence + $type-match DEFER to udd-check-lint, the composer that holds tokens+catalog+tree. A strict narrowing; the signature + the other 7 codes are UNCHANGED, no behavior added; §1 reject + §2 scenario amended to match. Fork A/D (human-approved at the freeze): the validator is NOT wired into cmd_check here — udd-check-lint LATER calls BOTH _catalog_tree_violations + _token_layer_violations inside cmd_check (named reds) plus the cross-file token resolution; cmd_check is NAMED-not-touched. Foundation shape = a NAMED SET — tokens.json · catalog.json · prototypes/<name>.json (Fork A). Ships templates/catalog.sample.json + templates/prototype.sample.json (a clean worked pair: Screen > Card > {Text, Text, Button}, token props bound to semantic tokens, a Button on-click passthrough) + templates/udd-catalog.md (the render recipe: a ~20-line catalog.json → json-render defineCatalog adapter; the flat tree feeds catalog.validate AS-IS), all ×3 byte-identical (test_bundle_parity + test_tree_parity). add.py-only logic change → ENGINE_MD5 c107344 → 774e025 → 3cdfaab (v3 — adversarial refute at VERIFY found a HARD-STOP crash + a structural fail-open, Option A close: the validator is now a pure TOTAL function that NEVER raises; a non-object component entry → malformed_catalog; a 9th code `malformed_element` flags a tree element / props / children of the wrong JSON type; an empty `children` array == absent. Strict fail-closed additions — the signature + the prior 8 codes' behavior are UNCHANGED; §1 reject + §2 scenarios amended; +3 red tests + the 4 refute-flagged weak tests strengthened). This pin supersedes re-aimed @ udd-token-schema (udd-design-foundation task 1/4: NEW pure validator _token_layer_violations(tokens) -> list[(code, path, detail)] + helper _token_value_form_ok — the compact-DTCG 3-layer token dialect's shape/citation checker. SIX named reds: unknown_layer · unknown_type · unresolved_alias · cross_layer_citation · primitive_has_alias · malformed_value. A token's LAYER is its top-level group name (primitive/semantic/component); fail-closed citation rule (primitive = literal only · semantic → primitive only · component → semantic only; aliases are the DTCG `{layer.dotted.path}` curly form, chains allowed, must terminate in a literal). Compact SCALAR value forms (color "#RRGGBB"/"#RRGGBBAA" · dimension "<n><unit>" · number · fontWeight 100-900|keyword · duration "<n>ms|s" · fontFamily str|[str]) DIVERGE from DTCG 2025.10's value objects — every divergence NAMED in templates/udd-tokens.md, which cites the Format Module 2025.10 Final CG Report. PURE (never mutates input), stdlib-only, DETERMINISTIC document order; [] == valid. Fork A (human-approved at the §3 freeze): the validator is NOT wired into cmd_check here — udd-check-lint LATER calls it inside cmd_check as named reds + adds the catalog/tree rules; cmd_check is NAMED-not-touched. Ships templates/tokens.sample.json (a clean 3-layer worked example) + templates/udd-tokens.md, both ×3 byte-identical (test_bundle_parity + test_tree_parity). add.py-only logic change → ENGINE_MD5 a4dcc0b → c107344 (8329cd4 was the pre-refute build; the verify-gate adversarial refute found a fail-OPEN gap — a token node carrying non-$ children skipped its subtree — so _index/_walk now ALWAYS descend, validating nested children fail-closed; guarded by test_nested_token_children_are_validated). This pin supersedes re-aimed @ gate-owner-marker (next-step-seams task 2/3: the footer + the `guide` TEXT next-step line now NAME the driver — ` [you drive]` (the AI proceeds) / ` [human gate]` (a human owns it) — filling the slot next-footer-engine reserved. ONE resolver, no second owner map: NEW _driver_stop(root, state, slug, phase) = _phase_owner(phase) != "ai" for EVERY phase EXCEPT verify, where it is _effective_autonomy(root, state, slug) != "auto" (the single dial seam — auto auto-gates → the AI; conservative/manual → the human; run.md:21 the contract freeze stays human dial-blind). _driver_marker(stop) renders the reserved word. _effective_autonomy reads the task's own `autonomy:` line, falling back to _project_autonomy when unset(None)/unknown("?") — never silently auto. _next_footer fills the slot: Arm A via _driver_stop(phase×autonomy); Arm B via the NEW _decide_next_pair(state, d) -> (text, human_stop) [the precedence extracted ONCE; _decide_next_base is now a thin [0] wrapper so the report rollup/digest callers are unchanged] — every decision point (decompose · HARD-STOP · goal-not-met · consolidate/archive · approve-contract · gate) → [human gate], the run-in-progress fallthrough → [you drive]; fail-soft carries NO marker (never assert a driver it could not compute). cmd_guide TEXT prints the marker on its `next   :` line (computed AFTER the unknown-phase guard, so a bad phase fails clean — the marker invents no default). Option F (human-confirmed at the freeze): the FROZEN machine-state-json JSON `owner`/`stop` is UNTOUCHED — changing it is a change-request against THAT contract (intake.md); the marker rides the footer + guide-TEXT only, and the guide-text↔guide-JSON divergence at verify-auto is a recorded §7 delta, not absorbed here. Sibling co-updates in test_next_footer_engine.py (§5 scope, both strengthenings): Arm B HARD-STOP exact-equal now pins ` [human gate]`; test_marker_slot_reserved→test_marker_slot_filled now asserts ` [you drive]`. This pin supersedes re-aimed @ next-footer-engine (whose ONE-engine-sourced-footer mechanics it preserves verbatim, described next), which itself carried — Arm A names the active in-flight task's phase command (advance, or the gate verbs at verify, guarded gate=="none" AND phase!="done" so a HARD-STOPped task falls to Arm B not back to re-gate) + its PHASE_GUIDE why; Arm B reuses _decide_next_base over the active-milestone rollup — the SAME precedence the report renders — with the empty-rows branch reshaped command-first ("decompose into tasks — add.py new-task <ms>", replacing the dead-end "none — no tasks yet"). Wired into advance·gate·new-task·new-milestone·phase·use·reopen·set-milestone·milestone-done·lock·stage·archive-milestone·compact, converging their ad-hoc tails (new-task "then: advance", new-milestone "Decompose…", gate HARD-STOP "return to BUILD", milestone-done "Confirm…") so each prints exactly ONE next-step. EXCEPTIONS (human-confirmed at tests, both keep their existing seam, no footer): the exit-3 HEAL paths (cmd_heal · gate→heal raise SystemExit(3) with a return_to_build redo message) and cmd_init (its bootstrap "next: run /add"/"lock" block already names the setup-flow next step; the generic resolver would regress it). Fail-soft: the footer is computed AFTER save_state and never crashes a saved mutation — no active milestone / unreadable doc / corrupt rollup degrade to "next: add.py status — re-orient". Pure render, writes nothing; the trailing marker slot is reserved for the sibling gate-owner-marker. read-only surfaces (status/guide/report) unchanged. _scope_guard no longer dies in place on the RECOVERABLE findings — an out-of-scope touch (source 'scope') and a present-but-wrong sidecar (diverged|unparseable, source 'scope-tamper') call the SAME _heal_or_escalate the tamper tripwire uses, returning the task to BUILD for an honest redo (exit 3) and counting against the ONE shared per-task HEAL_CAP; the (CAP+1)th confirmed finding records gate=HARD-STOP and escalates heal_exhausted. The ERASED baselines stay die-in-place (exit 1, no heal): a MISSING sidecar is scope_snapshot_tampered and an erased anchor under a present sidecar is scope_anchor_missing — a redo cannot recreate erased evidence (tripwire_missing parity). Every heal reason CARRIES its named code so the refusal-token assertions still match; placement unchanged — after _tamper_guard, BEFORE the waiver write, off the HARD-STOP path; a heal never rewrites the sidecar or anchor. check stays a read-only WARN monitor — never routes to heal, never increments. This pin supersedes re-aimed @ scope-violation-heal (whose recoverable-heal mechanics it preserves verbatim, described above). Carries the prior scope-gate-enforce gate (NEW _declared_scope parses the frozen §5 grammar — backticked tokens · FIRST 'Scope (may touch):' line · ./… task dir · '/' project root · bare sibling-of-previous · outside-root dropped · directory token = WHOLE subtree · no line = UNDECLARED grandfathered · garbage = [] no cover; tests->build _scope_walk {rel: md5} pruning _SCOPE_EXCLUDE_DIRS into sidecar scope-snapshot.json anchored by sidecar-md5; v2 anchor×sidecar co-witness after the anchor-erase REFUTE; v3 UNDECLARED-crossing cleanup), the engine-merge-base-enforcement wave-ledger gate (cmd_check wave section + read-only cmd_wave_verify, v4 status-field grammar, exactly-one echo column, six drift fixtures), and the engine-argv-portability re-bind. tool-agnostic — never runs git)) · re-aimed @ fold-suggestion-seams (surface the fold-into-PROJECT.md suggestion live in the loop: cmd_advance prints a per-task `add.py fold --task <slug>` hint at the OBSERVE transition where the loop lessons are captured; cmd_milestone_done previews the open lessons by competency + emits the ready-to-run fold command. Additive prose only — no decision path changes; the verb word is interpolated via _FOLD_VERB so the domain wording-lint sees no bare slang. ENGINE_MD5 fa8e9818 → 1911b250)
+ENGINE_PKG_MD5 = "001ad4a0f8c5e6a669c8bf0f3743f340"  # re-aimed @ extract-io-state (io_state.py joined the package). engine-package-skeleton: manifest digest over add_engine/*.py (engine_manifest.package_digest); pins the package modules so ENGINE_MD5 can stay md5(add.py) (the ~52 EnginePinTest copies rely on that). A LITERAL — engine_pin never hashes.
